@@ -1,9 +1,10 @@
 // @flow
 
-const DaemonFactory = require('ipfsd-ctl');
-const { default: AbortController } = require('abort-controller');
-const ipfsBin = require('go-ipfs-dep').path();
-const ipfsHttpModule = require('ipfs-http-client');
+import DaemonFactory from 'ipfsd-ctl';
+import { create } from 'ipfs-http-client';
+import { Agent } from 'http';
+import { path as ipfsBinPath } from 'go-ipfs';
+
 
 let nodes:Array<Object> = [];
 let pids: Array<number> = [];
@@ -12,13 +13,21 @@ let abortControllers: Array<AbortController> = [];
 const factory = DaemonFactory.createFactory({
   type: 'go',
   test: true,
-  ipfsBin,
-  ipfsHttpModule,
+  ipfsBin: ipfsBinPath(),
+  ipfsHttpModule: {
+    create: (url) => {
+      const options: Object = {
+        url,
+        agent: new Agent({ keepAlive: true, maxSockets: Infinity }),
+      };
+      return create(options);
+    },
+  },
   disposable: true,
   args: ['--enable-pubsub-experiment'],
 });
 
-module.exports.getGatewayIpfsNode = async (port:number) => {
+export const getGatewayIpfsNode = async (port:number) => {
   const options = {
     ipfsOptions: {
       config: {
@@ -34,6 +43,9 @@ module.exports.getGatewayIpfsNode = async (port:number) => {
             '/ip6/::/tcp/0',
           ],
         },
+        Pubsub: {
+          Enabled: true,
+        },
         Discovery: {
           MDNS: {
             Interval: 1,
@@ -44,23 +56,36 @@ module.exports.getGatewayIpfsNode = async (port:number) => {
     },
   };
   const daemon = await factory.spawn(options);
-  await daemon.api.id();
+  const { id } = await daemon.api.id();
+  daemon.api.ipfsId = id;
   const pid = await daemon.pid();
   pids.push(pid);
   return daemon.api;
 };
 
-module.exports.closeAllNodes = async () => {
+export const closeAllNodes = async () => {
   for (const abortController of abortControllers) {
     abortController.abort();
   }
   await factory.clean();
+  for (const pid of pids) {
+    try {
+      process.kill(pid, 'SIGINT');
+      await new Promise((resolve) => setTimeout(resolve, 100));
+      process.kill(pid, 'SIGINT');
+    } catch (error) {
+      if (error.code === 'ESRCH') {
+        continue;
+      }
+      throw error;
+    }
+  }
   nodes = [];
   abortControllers = [];
   pids = [];
 };
 
-const getIpfsNode = module.exports.getIpfsNode = async (bootstrap:Array<string> = []) => {
+export const getIpfsNode = async (bootstrap:Array<string> = []) => {
   const options = {
     ipfsOptions: {
       config: {
@@ -75,6 +100,9 @@ const getIpfsNode = module.exports.getIpfsNode = async (bootstrap:Array<string> 
             '/ip6/::/tcp/0',
           ],
         },
+        Pubsub: {
+          Enabled: true,
+        },
         Discovery: {
           MDNS: {
             Interval: 1,
@@ -84,14 +112,18 @@ const getIpfsNode = module.exports.getIpfsNode = async (bootstrap:Array<string> 
       },
     },
   };
+  const abortController = new AbortController();
+  abortControllers.push(abortController);
   const daemon = await factory.spawn(options);
-  await daemon.api.id();
+  const { id } = await daemon.api.id();
+  daemon.api.ipfsId = id;
   const pid = await daemon.pid();
   pids.push(pid);
+  daemon.api.abortControllerSignal = abortController.signal;
   return daemon.api;
 };
 
-module.exports.getSwarm = async (count:number) => {
+export const getSwarm = async (count:number) => {
   if (nodes.length < count) {
     nodes = nodes.concat(await Promise.all(Array.from({ length: count - nodes.length }, getIpfsNode)));
   }
@@ -101,7 +133,7 @@ module.exports.getSwarm = async (count:number) => {
   const abortController = new AbortController();
   abortControllers.push(abortController);
   for (const node of nodes) {
-    const { addresses } = await node.id({ signal: abortController.signal });
+    const { addresses } = await node.id();
     nodes.filter((x) => x !== node).forEach((x) => {
       x.swarm.connect(addresses[0], { signal: abortController.signal }).catch(() => {
         // This is a no-op because we are checking for the connected nodes in the next step
@@ -113,7 +145,7 @@ module.exports.getSwarm = async (count:number) => {
     await new Promise((resolve) => setTimeout(resolve, 1000));
     connectedNodes = 0;
     await Promise.all(nodes.map(async (node) => { // eslint-disable-line no-loop-func
-      const peers = await node.swarm.peers({ signal: abortController.signal });
+      const peers = await node.swarm.peers();
       if (peers.length >= nodes.length - 1) {
         connectedNodes += 1;
       }
