@@ -1,5 +1,5 @@
-// flow-typed signature: 4ad8afc0df7b9d1eefb6be4e99f44e8b
-// flow-typed version: <<STUB>>/codecov_v^3.0.0/flow_v0.63.1
+// flow-typed signature: 8f10056046c08189615c6daf37935d08
+// flow-typed version: <<STUB>>/codecov_v^3.8.2/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -42,11 +42,23 @@ declare module 'codecov/lib/services/appveyor' {
   declare module.exports: any;
 }
 
+declare module 'codecov/lib/services/azurePipelines' {
+  declare module.exports: any;
+}
+
 declare module 'codecov/lib/services/buildkite' {
   declare module.exports: any;
 }
 
 declare module 'codecov/lib/services/circle' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/lib/services/cirrus' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/lib/services/codebuild' {
   declare module.exports: any;
 }
 
@@ -58,7 +70,15 @@ declare module 'codecov/lib/services/drone' {
   declare module.exports: any;
 }
 
+declare module 'codecov/lib/services/github_actions' {
+  declare module.exports: any;
+}
+
 declare module 'codecov/lib/services/gitlab' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/lib/services/heroku' {
   declare module.exports: any;
 }
 
@@ -74,11 +94,19 @@ declare module 'codecov/lib/services/semaphore' {
   declare module.exports: any;
 }
 
+declare module 'codecov/lib/services/semaphore2x' {
+  declare module.exports: any;
+}
+
 declare module 'codecov/lib/services/shippable' {
   declare module.exports: any;
 }
 
 declare module 'codecov/lib/services/snap' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/lib/services/teamcity' {
   declare module.exports: any;
 }
 
@@ -90,75 +118,99 @@ declare module 'codecov/lib/services/wercker' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/detect' {
+declare module 'codecov/test/detect.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/git' {
+declare module 'codecov/test/git.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/index' {
+declare module 'codecov/test/index.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/appveyor' {
+declare module 'codecov/test/services/appveyor.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/buildkite' {
+declare module 'codecov/test/services/azure-pipelines.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/circle' {
+declare module 'codecov/test/services/buildkite.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/codeship' {
+declare module 'codecov/test/services/circle.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/drone' {
+declare module 'codecov/test/services/cirrus.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/gitlab' {
+declare module 'codecov/test/services/codebuild.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/jenkins' {
+declare module 'codecov/test/services/codeship.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/localGit' {
+declare module 'codecov/test/services/drone.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/semaphore' {
+declare module 'codecov/test/services/github_actions.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/shippable' {
+declare module 'codecov/test/services/gitlab.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/snap' {
+declare module 'codecov/test/services/heroku.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/travis' {
+declare module 'codecov/test/services/jenkins.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/services/wercker' {
+declare module 'codecov/test/services/localGit.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/test/upload' {
+declare module 'codecov/test/services/semaphore.test' {
   declare module.exports: any;
 }
 
-declare module 'codecov/testinit' {
+declare module 'codecov/test/services/semaphore2x.test' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/test/services/shippable.test' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/test/services/snap.test' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/test/services/teamcity.test' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/test/services/travis.test' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/test/services/wercker.test' {
+  declare module.exports: any;
+}
+
+declare module 'codecov/test/upload.test' {
   declare module.exports: any;
 }
 
@@ -184,11 +236,20 @@ declare module 'codecov/lib/offline.js' {
 declare module 'codecov/lib/services/appveyor.js' {
   declare module.exports: $Exports<'codecov/lib/services/appveyor'>;
 }
+declare module 'codecov/lib/services/azurePipelines.js' {
+  declare module.exports: $Exports<'codecov/lib/services/azurePipelines'>;
+}
 declare module 'codecov/lib/services/buildkite.js' {
   declare module.exports: $Exports<'codecov/lib/services/buildkite'>;
 }
 declare module 'codecov/lib/services/circle.js' {
   declare module.exports: $Exports<'codecov/lib/services/circle'>;
+}
+declare module 'codecov/lib/services/cirrus.js' {
+  declare module.exports: $Exports<'codecov/lib/services/cirrus'>;
+}
+declare module 'codecov/lib/services/codebuild.js' {
+  declare module.exports: $Exports<'codecov/lib/services/codebuild'>;
 }
 declare module 'codecov/lib/services/codeship.js' {
   declare module.exports: $Exports<'codecov/lib/services/codeship'>;
@@ -196,8 +257,14 @@ declare module 'codecov/lib/services/codeship.js' {
 declare module 'codecov/lib/services/drone.js' {
   declare module.exports: $Exports<'codecov/lib/services/drone'>;
 }
+declare module 'codecov/lib/services/github_actions.js' {
+  declare module.exports: $Exports<'codecov/lib/services/github_actions'>;
+}
 declare module 'codecov/lib/services/gitlab.js' {
   declare module.exports: $Exports<'codecov/lib/services/gitlab'>;
+}
+declare module 'codecov/lib/services/heroku.js' {
+  declare module.exports: $Exports<'codecov/lib/services/heroku'>;
 }
 declare module 'codecov/lib/services/jenkins.js' {
   declare module.exports: $Exports<'codecov/lib/services/jenkins'>;
@@ -208,11 +275,17 @@ declare module 'codecov/lib/services/localGit.js' {
 declare module 'codecov/lib/services/semaphore.js' {
   declare module.exports: $Exports<'codecov/lib/services/semaphore'>;
 }
+declare module 'codecov/lib/services/semaphore2x.js' {
+  declare module.exports: $Exports<'codecov/lib/services/semaphore2x'>;
+}
 declare module 'codecov/lib/services/shippable.js' {
   declare module.exports: $Exports<'codecov/lib/services/shippable'>;
 }
 declare module 'codecov/lib/services/snap.js' {
   declare module.exports: $Exports<'codecov/lib/services/snap'>;
+}
+declare module 'codecov/lib/services/teamcity.js' {
+  declare module.exports: $Exports<'codecov/lib/services/teamcity'>;
 }
 declare module 'codecov/lib/services/travis.js' {
   declare module.exports: $Exports<'codecov/lib/services/travis'>;
@@ -220,57 +293,75 @@ declare module 'codecov/lib/services/travis.js' {
 declare module 'codecov/lib/services/wercker.js' {
   declare module.exports: $Exports<'codecov/lib/services/wercker'>;
 }
-declare module 'codecov/test/detect.js' {
-  declare module.exports: $Exports<'codecov/test/detect'>;
+declare module 'codecov/test/detect.test.js' {
+  declare module.exports: $Exports<'codecov/test/detect.test'>;
 }
-declare module 'codecov/test/git.js' {
-  declare module.exports: $Exports<'codecov/test/git'>;
+declare module 'codecov/test/git.test.js' {
+  declare module.exports: $Exports<'codecov/test/git.test'>;
 }
-declare module 'codecov/test/index.js' {
-  declare module.exports: $Exports<'codecov/test/index'>;
+declare module 'codecov/test/index.test.js' {
+  declare module.exports: $Exports<'codecov/test/index.test'>;
 }
-declare module 'codecov/test/services/appveyor.js' {
-  declare module.exports: $Exports<'codecov/test/services/appveyor'>;
+declare module 'codecov/test/services/appveyor.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/appveyor.test'>;
 }
-declare module 'codecov/test/services/buildkite.js' {
-  declare module.exports: $Exports<'codecov/test/services/buildkite'>;
+declare module 'codecov/test/services/azure-pipelines.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/azure-pipelines.test'>;
 }
-declare module 'codecov/test/services/circle.js' {
-  declare module.exports: $Exports<'codecov/test/services/circle'>;
+declare module 'codecov/test/services/buildkite.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/buildkite.test'>;
 }
-declare module 'codecov/test/services/codeship.js' {
-  declare module.exports: $Exports<'codecov/test/services/codeship'>;
+declare module 'codecov/test/services/circle.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/circle.test'>;
 }
-declare module 'codecov/test/services/drone.js' {
-  declare module.exports: $Exports<'codecov/test/services/drone'>;
+declare module 'codecov/test/services/cirrus.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/cirrus.test'>;
 }
-declare module 'codecov/test/services/gitlab.js' {
-  declare module.exports: $Exports<'codecov/test/services/gitlab'>;
+declare module 'codecov/test/services/codebuild.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/codebuild.test'>;
 }
-declare module 'codecov/test/services/jenkins.js' {
-  declare module.exports: $Exports<'codecov/test/services/jenkins'>;
+declare module 'codecov/test/services/codeship.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/codeship.test'>;
 }
-declare module 'codecov/test/services/localGit.js' {
-  declare module.exports: $Exports<'codecov/test/services/localGit'>;
+declare module 'codecov/test/services/drone.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/drone.test'>;
 }
-declare module 'codecov/test/services/semaphore.js' {
-  declare module.exports: $Exports<'codecov/test/services/semaphore'>;
+declare module 'codecov/test/services/github_actions.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/github_actions.test'>;
 }
-declare module 'codecov/test/services/shippable.js' {
-  declare module.exports: $Exports<'codecov/test/services/shippable'>;
+declare module 'codecov/test/services/gitlab.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/gitlab.test'>;
 }
-declare module 'codecov/test/services/snap.js' {
-  declare module.exports: $Exports<'codecov/test/services/snap'>;
+declare module 'codecov/test/services/heroku.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/heroku.test'>;
 }
-declare module 'codecov/test/services/travis.js' {
-  declare module.exports: $Exports<'codecov/test/services/travis'>;
+declare module 'codecov/test/services/jenkins.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/jenkins.test'>;
 }
-declare module 'codecov/test/services/wercker.js' {
-  declare module.exports: $Exports<'codecov/test/services/wercker'>;
+declare module 'codecov/test/services/localGit.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/localGit.test'>;
 }
-declare module 'codecov/test/upload.js' {
-  declare module.exports: $Exports<'codecov/test/upload'>;
+declare module 'codecov/test/services/semaphore.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/semaphore.test'>;
 }
-declare module 'codecov/testinit.js' {
-  declare module.exports: $Exports<'codecov/testinit'>;
+declare module 'codecov/test/services/semaphore2x.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/semaphore2x.test'>;
+}
+declare module 'codecov/test/services/shippable.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/shippable.test'>;
+}
+declare module 'codecov/test/services/snap.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/snap.test'>;
+}
+declare module 'codecov/test/services/teamcity.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/teamcity.test'>;
+}
+declare module 'codecov/test/services/travis.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/travis.test'>;
+}
+declare module 'codecov/test/services/wercker.test.js' {
+  declare module.exports: $Exports<'codecov/test/services/wercker.test'>;
+}
+declare module 'codecov/test/upload.test.js' {
+  declare module.exports: $Exports<'codecov/test/upload.test'>;
 }
