@@ -1,5 +1,5 @@
-// flow-typed signature: bc75bd8142e253c635297db7a8f079e7
-// flow-typed version: <<STUB>>/cross-env_v^5.1.3/flow_v0.63.1
+// flow-typed signature: 6a537b9df2fbc5cfda4915875f489c13
+// flow-typed version: <<STUB>>/cross-env_v^7.0.3/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,39 +22,49 @@ declare module 'cross-env' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'cross-env/dist/bin/cross-env-shell' {
+declare module 'cross-env/src/bin/cross-env-shell' {
   declare module.exports: any;
 }
 
-declare module 'cross-env/dist/bin/cross-env' {
+declare module 'cross-env/src/bin/cross-env' {
   declare module.exports: any;
 }
 
-declare module 'cross-env/dist/command' {
+declare module 'cross-env/src/command' {
   declare module.exports: any;
 }
 
-declare module 'cross-env/dist/index' {
+declare module 'cross-env/src' {
   declare module.exports: any;
 }
 
-declare module 'cross-env/dist/variable' {
+declare module 'cross-env/src/is-windows' {
+  declare module.exports: any;
+}
+
+declare module 'cross-env/src/variable' {
   declare module.exports: any;
 }
 
 // Filename aliases
-declare module 'cross-env/dist/bin/cross-env-shell.js' {
-  declare module.exports: $Exports<'cross-env/dist/bin/cross-env-shell'>;
+declare module 'cross-env/src/bin/cross-env-shell.js' {
+  declare module.exports: $Exports<'cross-env/src/bin/cross-env-shell'>;
 }
-declare module 'cross-env/dist/bin/cross-env.js' {
-  declare module.exports: $Exports<'cross-env/dist/bin/cross-env'>;
+declare module 'cross-env/src/bin/cross-env.js' {
+  declare module.exports: $Exports<'cross-env/src/bin/cross-env'>;
 }
-declare module 'cross-env/dist/command.js' {
-  declare module.exports: $Exports<'cross-env/dist/command'>;
+declare module 'cross-env/src/command.js' {
+  declare module.exports: $Exports<'cross-env/src/command'>;
 }
-declare module 'cross-env/dist/index.js' {
-  declare module.exports: $Exports<'cross-env/dist/index'>;
+declare module 'cross-env/src/index' {
+  declare module.exports: $Exports<'cross-env/src'>;
 }
-declare module 'cross-env/dist/variable.js' {
-  declare module.exports: $Exports<'cross-env/dist/variable'>;
+declare module 'cross-env/src/index.js' {
+  declare module.exports: $Exports<'cross-env/src'>;
+}
+declare module 'cross-env/src/is-windows.js' {
+  declare module.exports: $Exports<'cross-env/src/is-windows'>;
+}
+declare module 'cross-env/src/variable.js' {
+  declare module.exports: $Exports<'cross-env/src/variable'>;
 }
