@@ -1,5 +1,5 @@
-// flow-typed signature: e470e104d72f82f78e7f55c5adb33536
-// flow-typed version: <<STUB>>/documentation_v^5.3.5/flow_v0.63.1
+// flow-typed signature: 2623e6269973a96dc22c80459934d955
+// flow-typed version: <<STUB>>/documentation_v^13.2.5/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,398 +22,6 @@ declare module 'documentation' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'documentation/__tests__/bin-readme' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/bin-watch-serve' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/bin' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/auto_lang_hljs/multilanguage.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/boolean-literal-type.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/class.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/custom_theme/index' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/default-export-function.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported-bad/exports-z' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported-bad/x' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported-export-default-object.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported-export-default-value.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported/x' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported/y' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/document-exported/z' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/es6-class.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/es6-default2.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/es6-import.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/es6.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/event.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/example-caption.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/extension' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/extension/extension-required' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/external.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/factory.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/flow-unnamed-params.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/html/nested.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/infer-private.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/inheritance.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/inline-link.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/interface.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/internal.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/lends.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/lint/lint.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/literal_types.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/memberedclass.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/merge-infered-type.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/meta.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/multisignature.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/nearby_params.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/nest_params.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/newline-in-description.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/no-name.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/optional-record-field-type.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/params.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/react-jsx.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/readme/index' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/require-json-no-extension.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/require-json.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/sections.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple-callback.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple-hashbang.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple-private.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple-singlestar.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple-triplestar.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple-two.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/simple.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/sort-order-alpha.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/sorting/input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/string-literal-key.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/system-import.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/this-class.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/type_application.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/fixture/var-function-param-return.input' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/format_type' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/index' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/filter_access' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/flow_doctrine' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/git/find_git' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/git/url_prefix' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/github' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/hierarchy' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/access' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/augments' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/finders' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/kind' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/membership' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/name' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/params' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/properties' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/return' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/infer/type' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/input/dependency' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/input/shallow' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/lint' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/merge_config' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/module_filters' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/nest' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/output/util/formatters' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/parse' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/parsers/javascript' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/server' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/sort' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/lib/walk' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/linker' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/test' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/__tests__/utils' {
-  declare module.exports: any;
-}
-
 declare module 'documentation/bin/documentation' {
   declare module.exports: any;
 }
@@ -422,235 +30,11 @@ declare module 'documentation/declarations/comment' {
   declare module.exports: any;
 }
 
-declare module 'documentation/lib/commands/build' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/commands/index' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/commands/lint' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/commands/readme' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/commands/serve' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/commands/shared_options' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/default_theme/assets/anchor' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/default_theme/assets/site' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/default_theme/assets/split' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/default_theme/index' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/extractors/comments' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/extractors/exported' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/filter_access' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/flow_doctrine' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/garbage_collect' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/git/find_git' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/git/url_prefix' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/github' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/hierarchy' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/index' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/access' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/augments' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/finders' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/kind' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/membership' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/name' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/params' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/properties' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/return' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/infer/type' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/inline_tokenizer' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/input/dependency' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/input/shallow' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/is_jsdoc_comment' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/lint' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/merge_config' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/module_filters' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/nest' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/highlighter' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/html' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/json' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/markdown_ast' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/markdown' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/util/format_type' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/util/formatters' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/util/linker_stack' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/output/util/reroute_links' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/parse_markdown' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/parse' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/parsers/javascript' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/parsers/parse_to_ast' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/serve/error_page' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/serve/server' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/smart_glob' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/sort' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/lib/walk' {
-  declare module.exports: any;
-}
-
 declare module 'documentation/src/commands/build' {
   declare module.exports: any;
 }
 
-declare module 'documentation/src/commands/index' {
+declare module 'documentation/src/commands' {
   declare module.exports: any;
 }
 
@@ -682,7 +66,7 @@ declare module 'documentation/src/default_theme/assets/split' {
   declare module.exports: any;
 }
 
-declare module 'documentation/src/default_theme/index' {
+declare module 'documentation/src/default_theme' {
   declare module.exports: any;
 }
 
@@ -706,6 +90,10 @@ declare module 'documentation/src/garbage_collect' {
   declare module.exports: any;
 }
 
+declare module 'documentation/src/get-readme-file' {
+  declare module.exports: any;
+}
+
 declare module 'documentation/src/git/find_git' {
   declare module.exports: any;
 }
@@ -722,7 +110,7 @@ declare module 'documentation/src/hierarchy' {
   declare module.exports: any;
 }
 
-declare module 'documentation/src/index' {
+declare module 'documentation/src' {
   declare module.exports: any;
 }
 
@@ -735,6 +123,10 @@ declare module 'documentation/src/infer/augments' {
 }
 
 declare module 'documentation/src/infer/finders' {
+  declare module.exports: any;
+}
+
+declare module 'documentation/src/infer/implements' {
   declare module.exports: any;
 }
 
@@ -763,10 +155,6 @@ declare module 'documentation/src/infer/return' {
 }
 
 declare module 'documentation/src/infer/type' {
-  declare module.exports: any;
-}
-
-declare module 'documentation/src/inline_tokenizer' {
   declare module.exports: any;
 }
 
@@ -834,10 +222,6 @@ declare module 'documentation/src/output/util/reroute_links' {
   declare module.exports: any;
 }
 
-declare module 'documentation/src/parse_markdown' {
-  declare module.exports: any;
-}
-
 declare module 'documentation/src/parse' {
   declare module.exports: any;
 }
@@ -847,6 +231,22 @@ declare module 'documentation/src/parsers/javascript' {
 }
 
 declare module 'documentation/src/parsers/parse_to_ast' {
+  declare module.exports: any;
+}
+
+declare module 'documentation/src/parsers/vue' {
+  declare module.exports: any;
+}
+
+declare module 'documentation/src/remark-jsDoc-link' {
+  declare module.exports: any;
+}
+
+declare module 'documentation/src/remark-parse' {
+  declare module.exports: any;
+}
+
+declare module 'documentation/src/remark-remove-position' {
   declare module.exports: any;
 }
 
@@ -866,484 +266,33 @@ declare module 'documentation/src/sort' {
   declare module.exports: any;
 }
 
+declare module 'documentation/src/ts_doctrine' {
+  declare module.exports: any;
+}
+
+declare module 'documentation/src/type_annotation' {
+  declare module.exports: any;
+}
+
 declare module 'documentation/src/walk' {
   declare module.exports: any;
 }
 
 // Filename aliases
-declare module 'documentation/__tests__/bin-readme.js' {
-  declare module.exports: $Exports<'documentation/__tests__/bin-readme'>;
-}
-declare module 'documentation/__tests__/bin-watch-serve.js' {
-  declare module.exports: $Exports<'documentation/__tests__/bin-watch-serve'>;
-}
-declare module 'documentation/__tests__/bin.js' {
-  declare module.exports: $Exports<'documentation/__tests__/bin'>;
-}
-declare module 'documentation/__tests__/fixture/auto_lang_hljs/multilanguage.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/auto_lang_hljs/multilanguage.input'>;
-}
-declare module 'documentation/__tests__/fixture/boolean-literal-type.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/boolean-literal-type.input'>;
-}
-declare module 'documentation/__tests__/fixture/class.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/class.input'>;
-}
-declare module 'documentation/__tests__/fixture/custom_theme/index.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/custom_theme/index'>;
-}
-declare module 'documentation/__tests__/fixture/default-export-function.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/default-export-function.input'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported-bad/exports-z.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported-bad/exports-z'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported-bad/x.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported-bad/x'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported-export-default-object.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported-export-default-object.input'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported-export-default-value.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported-export-default-value.input'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported.input'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported/x.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported/x'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported/y.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported/y'>;
-}
-declare module 'documentation/__tests__/fixture/document-exported/z.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/document-exported/z'>;
-}
-declare module 'documentation/__tests__/fixture/es6-class.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/es6-class.input'>;
-}
-declare module 'documentation/__tests__/fixture/es6-default2.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/es6-default2.input'>;
-}
-declare module 'documentation/__tests__/fixture/es6-import.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/es6-import.input'>;
-}
-declare module 'documentation/__tests__/fixture/es6.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/es6.input'>;
-}
-declare module 'documentation/__tests__/fixture/event.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/event.input'>;
-}
-declare module 'documentation/__tests__/fixture/example-caption.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/example-caption.input'>;
-}
-declare module 'documentation/__tests__/fixture/extension.jsx' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/extension'>;
-}
-declare module 'documentation/__tests__/fixture/extension/extension-required.jsx' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/extension/extension-required'>;
-}
-declare module 'documentation/__tests__/fixture/external.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/external.input'>;
-}
-declare module 'documentation/__tests__/fixture/factory.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/factory.input'>;
-}
-declare module 'documentation/__tests__/fixture/flow-unnamed-params.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/flow-unnamed-params.input'>;
-}
-declare module 'documentation/__tests__/fixture/html/nested.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/html/nested.input'>;
-}
-declare module 'documentation/__tests__/fixture/infer-private.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/infer-private.input'>;
-}
-declare module 'documentation/__tests__/fixture/inheritance.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/inheritance.input'>;
-}
-declare module 'documentation/__tests__/fixture/inline-link.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/inline-link.input'>;
-}
-declare module 'documentation/__tests__/fixture/interface.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/interface.input'>;
-}
-declare module 'documentation/__tests__/fixture/internal.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/internal.input'>;
-}
-declare module 'documentation/__tests__/fixture/lends.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/lends.input'>;
-}
-declare module 'documentation/__tests__/fixture/lint/lint.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/lint/lint.input'>;
-}
-declare module 'documentation/__tests__/fixture/literal_types.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/literal_types.input'>;
-}
-declare module 'documentation/__tests__/fixture/memberedclass.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/memberedclass.input'>;
-}
-declare module 'documentation/__tests__/fixture/merge-infered-type.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/merge-infered-type.input'>;
-}
-declare module 'documentation/__tests__/fixture/meta.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/meta.input'>;
-}
-declare module 'documentation/__tests__/fixture/multisignature.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/multisignature.input'>;
-}
-declare module 'documentation/__tests__/fixture/nearby_params.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/nearby_params.input'>;
-}
-declare module 'documentation/__tests__/fixture/nest_params.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/nest_params.input'>;
-}
-declare module 'documentation/__tests__/fixture/newline-in-description.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/newline-in-description.input'>;
-}
-declare module 'documentation/__tests__/fixture/no-name.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/no-name.input'>;
-}
-declare module 'documentation/__tests__/fixture/optional-record-field-type.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/optional-record-field-type.input'>;
-}
-declare module 'documentation/__tests__/fixture/params.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/params.input'>;
-}
-declare module 'documentation/__tests__/fixture/react-jsx.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/react-jsx.input'>;
-}
-declare module 'documentation/__tests__/fixture/readme/index.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/readme/index'>;
-}
-declare module 'documentation/__tests__/fixture/require-json-no-extension.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/require-json-no-extension.input'>;
-}
-declare module 'documentation/__tests__/fixture/require-json.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/require-json.input'>;
-}
-declare module 'documentation/__tests__/fixture/sections.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/sections.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple-callback.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple-callback.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple-hashbang.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple-hashbang.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple-private.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple-private.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple-singlestar.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple-singlestar.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple-triplestar.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple-triplestar.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple-two.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple-two.input'>;
-}
-declare module 'documentation/__tests__/fixture/simple.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/simple.input'>;
-}
-declare module 'documentation/__tests__/fixture/sort-order-alpha.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/sort-order-alpha.input'>;
-}
-declare module 'documentation/__tests__/fixture/sorting/input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/sorting/input'>;
-}
-declare module 'documentation/__tests__/fixture/string-literal-key.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/string-literal-key.input'>;
-}
-declare module 'documentation/__tests__/fixture/system-import.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/system-import.input'>;
-}
-declare module 'documentation/__tests__/fixture/this-class.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/this-class.input'>;
-}
-declare module 'documentation/__tests__/fixture/type_application.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/type_application.input'>;
-}
-declare module 'documentation/__tests__/fixture/var-function-param-return.input.js' {
-  declare module.exports: $Exports<'documentation/__tests__/fixture/var-function-param-return.input'>;
-}
-declare module 'documentation/__tests__/format_type.js' {
-  declare module.exports: $Exports<'documentation/__tests__/format_type'>;
-}
-declare module 'documentation/__tests__/index.js' {
-  declare module.exports: $Exports<'documentation/__tests__/index'>;
-}
-declare module 'documentation/__tests__/lib/filter_access.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/filter_access'>;
-}
-declare module 'documentation/__tests__/lib/flow_doctrine.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/flow_doctrine'>;
-}
-declare module 'documentation/__tests__/lib/git/find_git.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/git/find_git'>;
-}
-declare module 'documentation/__tests__/lib/git/url_prefix.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/git/url_prefix'>;
-}
-declare module 'documentation/__tests__/lib/github.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/github'>;
-}
-declare module 'documentation/__tests__/lib/hierarchy.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/hierarchy'>;
-}
-declare module 'documentation/__tests__/lib/infer/access.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/access'>;
-}
-declare module 'documentation/__tests__/lib/infer/augments.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/augments'>;
-}
-declare module 'documentation/__tests__/lib/infer/finders.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/finders'>;
-}
-declare module 'documentation/__tests__/lib/infer/kind.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/kind'>;
-}
-declare module 'documentation/__tests__/lib/infer/membership.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/membership'>;
-}
-declare module 'documentation/__tests__/lib/infer/name.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/name'>;
-}
-declare module 'documentation/__tests__/lib/infer/params.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/params'>;
-}
-declare module 'documentation/__tests__/lib/infer/properties.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/properties'>;
-}
-declare module 'documentation/__tests__/lib/infer/return.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/return'>;
-}
-declare module 'documentation/__tests__/lib/infer/type.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/infer/type'>;
-}
-declare module 'documentation/__tests__/lib/input/dependency.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/input/dependency'>;
-}
-declare module 'documentation/__tests__/lib/input/shallow.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/input/shallow'>;
-}
-declare module 'documentation/__tests__/lib/lint.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/lint'>;
-}
-declare module 'documentation/__tests__/lib/merge_config.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/merge_config'>;
-}
-declare module 'documentation/__tests__/lib/module_filters.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/module_filters'>;
-}
-declare module 'documentation/__tests__/lib/nest.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/nest'>;
-}
-declare module 'documentation/__tests__/lib/output/util/formatters.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/output/util/formatters'>;
-}
-declare module 'documentation/__tests__/lib/parse.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/parse'>;
-}
-declare module 'documentation/__tests__/lib/parsers/javascript.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/parsers/javascript'>;
-}
-declare module 'documentation/__tests__/lib/server.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/server'>;
-}
-declare module 'documentation/__tests__/lib/sort.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/sort'>;
-}
-declare module 'documentation/__tests__/lib/walk.js' {
-  declare module.exports: $Exports<'documentation/__tests__/lib/walk'>;
-}
-declare module 'documentation/__tests__/linker.js' {
-  declare module.exports: $Exports<'documentation/__tests__/linker'>;
-}
-declare module 'documentation/__tests__/test.js' {
-  declare module.exports: $Exports<'documentation/__tests__/test'>;
-}
-declare module 'documentation/__tests__/utils.js' {
-  declare module.exports: $Exports<'documentation/__tests__/utils'>;
-}
 declare module 'documentation/bin/documentation.js' {
   declare module.exports: $Exports<'documentation/bin/documentation'>;
 }
 declare module 'documentation/declarations/comment.js' {
   declare module.exports: $Exports<'documentation/declarations/comment'>;
 }
-declare module 'documentation/lib/commands/build.js' {
-  declare module.exports: $Exports<'documentation/lib/commands/build'>;
-}
-declare module 'documentation/lib/commands/index.js' {
-  declare module.exports: $Exports<'documentation/lib/commands/index'>;
-}
-declare module 'documentation/lib/commands/lint.js' {
-  declare module.exports: $Exports<'documentation/lib/commands/lint'>;
-}
-declare module 'documentation/lib/commands/readme.js' {
-  declare module.exports: $Exports<'documentation/lib/commands/readme'>;
-}
-declare module 'documentation/lib/commands/serve.js' {
-  declare module.exports: $Exports<'documentation/lib/commands/serve'>;
-}
-declare module 'documentation/lib/commands/shared_options.js' {
-  declare module.exports: $Exports<'documentation/lib/commands/shared_options'>;
-}
-declare module 'documentation/lib/default_theme/assets/anchor.js' {
-  declare module.exports: $Exports<'documentation/lib/default_theme/assets/anchor'>;
-}
-declare module 'documentation/lib/default_theme/assets/site.js' {
-  declare module.exports: $Exports<'documentation/lib/default_theme/assets/site'>;
-}
-declare module 'documentation/lib/default_theme/assets/split.js' {
-  declare module.exports: $Exports<'documentation/lib/default_theme/assets/split'>;
-}
-declare module 'documentation/lib/default_theme/index.js' {
-  declare module.exports: $Exports<'documentation/lib/default_theme/index'>;
-}
-declare module 'documentation/lib/extractors/comments.js' {
-  declare module.exports: $Exports<'documentation/lib/extractors/comments'>;
-}
-declare module 'documentation/lib/extractors/exported.js' {
-  declare module.exports: $Exports<'documentation/lib/extractors/exported'>;
-}
-declare module 'documentation/lib/filter_access.js' {
-  declare module.exports: $Exports<'documentation/lib/filter_access'>;
-}
-declare module 'documentation/lib/flow_doctrine.js' {
-  declare module.exports: $Exports<'documentation/lib/flow_doctrine'>;
-}
-declare module 'documentation/lib/garbage_collect.js' {
-  declare module.exports: $Exports<'documentation/lib/garbage_collect'>;
-}
-declare module 'documentation/lib/git/find_git.js' {
-  declare module.exports: $Exports<'documentation/lib/git/find_git'>;
-}
-declare module 'documentation/lib/git/url_prefix.js' {
-  declare module.exports: $Exports<'documentation/lib/git/url_prefix'>;
-}
-declare module 'documentation/lib/github.js' {
-  declare module.exports: $Exports<'documentation/lib/github'>;
-}
-declare module 'documentation/lib/hierarchy.js' {
-  declare module.exports: $Exports<'documentation/lib/hierarchy'>;
-}
-declare module 'documentation/lib/index.js' {
-  declare module.exports: $Exports<'documentation/lib/index'>;
-}
-declare module 'documentation/lib/infer/access.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/access'>;
-}
-declare module 'documentation/lib/infer/augments.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/augments'>;
-}
-declare module 'documentation/lib/infer/finders.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/finders'>;
-}
-declare module 'documentation/lib/infer/kind.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/kind'>;
-}
-declare module 'documentation/lib/infer/membership.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/membership'>;
-}
-declare module 'documentation/lib/infer/name.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/name'>;
-}
-declare module 'documentation/lib/infer/params.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/params'>;
-}
-declare module 'documentation/lib/infer/properties.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/properties'>;
-}
-declare module 'documentation/lib/infer/return.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/return'>;
-}
-declare module 'documentation/lib/infer/type.js' {
-  declare module.exports: $Exports<'documentation/lib/infer/type'>;
-}
-declare module 'documentation/lib/inline_tokenizer.js' {
-  declare module.exports: $Exports<'documentation/lib/inline_tokenizer'>;
-}
-declare module 'documentation/lib/input/dependency.js' {
-  declare module.exports: $Exports<'documentation/lib/input/dependency'>;
-}
-declare module 'documentation/lib/input/shallow.js' {
-  declare module.exports: $Exports<'documentation/lib/input/shallow'>;
-}
-declare module 'documentation/lib/is_jsdoc_comment.js' {
-  declare module.exports: $Exports<'documentation/lib/is_jsdoc_comment'>;
-}
-declare module 'documentation/lib/lint.js' {
-  declare module.exports: $Exports<'documentation/lib/lint'>;
-}
-declare module 'documentation/lib/merge_config.js' {
-  declare module.exports: $Exports<'documentation/lib/merge_config'>;
-}
-declare module 'documentation/lib/module_filters.js' {
-  declare module.exports: $Exports<'documentation/lib/module_filters'>;
-}
-declare module 'documentation/lib/nest.js' {
-  declare module.exports: $Exports<'documentation/lib/nest'>;
-}
-declare module 'documentation/lib/output/highlighter.js' {
-  declare module.exports: $Exports<'documentation/lib/output/highlighter'>;
-}
-declare module 'documentation/lib/output/html.js' {
-  declare module.exports: $Exports<'documentation/lib/output/html'>;
-}
-declare module 'documentation/lib/output/json.js' {
-  declare module.exports: $Exports<'documentation/lib/output/json'>;
-}
-declare module 'documentation/lib/output/markdown_ast.js' {
-  declare module.exports: $Exports<'documentation/lib/output/markdown_ast'>;
-}
-declare module 'documentation/lib/output/markdown.js' {
-  declare module.exports: $Exports<'documentation/lib/output/markdown'>;
-}
-declare module 'documentation/lib/output/util/format_type.js' {
-  declare module.exports: $Exports<'documentation/lib/output/util/format_type'>;
-}
-declare module 'documentation/lib/output/util/formatters.js' {
-  declare module.exports: $Exports<'documentation/lib/output/util/formatters'>;
-}
-declare module 'documentation/lib/output/util/linker_stack.js' {
-  declare module.exports: $Exports<'documentation/lib/output/util/linker_stack'>;
-}
-declare module 'documentation/lib/output/util/reroute_links.js' {
-  declare module.exports: $Exports<'documentation/lib/output/util/reroute_links'>;
-}
-declare module 'documentation/lib/parse_markdown.js' {
-  declare module.exports: $Exports<'documentation/lib/parse_markdown'>;
-}
-declare module 'documentation/lib/parse.js' {
-  declare module.exports: $Exports<'documentation/lib/parse'>;
-}
-declare module 'documentation/lib/parsers/javascript.js' {
-  declare module.exports: $Exports<'documentation/lib/parsers/javascript'>;
-}
-declare module 'documentation/lib/parsers/parse_to_ast.js' {
-  declare module.exports: $Exports<'documentation/lib/parsers/parse_to_ast'>;
-}
-declare module 'documentation/lib/serve/error_page.js' {
-  declare module.exports: $Exports<'documentation/lib/serve/error_page'>;
-}
-declare module 'documentation/lib/serve/server.js' {
-  declare module.exports: $Exports<'documentation/lib/serve/server'>;
-}
-declare module 'documentation/lib/smart_glob.js' {
-  declare module.exports: $Exports<'documentation/lib/smart_glob'>;
-}
-declare module 'documentation/lib/sort.js' {
-  declare module.exports: $Exports<'documentation/lib/sort'>;
-}
-declare module 'documentation/lib/walk.js' {
-  declare module.exports: $Exports<'documentation/lib/walk'>;
-}
 declare module 'documentation/src/commands/build.js' {
   declare module.exports: $Exports<'documentation/src/commands/build'>;
 }
+declare module 'documentation/src/commands/index' {
+  declare module.exports: $Exports<'documentation/src/commands'>;
+}
 declare module 'documentation/src/commands/index.js' {
-  declare module.exports: $Exports<'documentation/src/commands/index'>;
+  declare module.exports: $Exports<'documentation/src/commands'>;
 }
 declare module 'documentation/src/commands/lint.js' {
   declare module.exports: $Exports<'documentation/src/commands/lint'>;
@@ -1366,8 +315,11 @@ declare module 'documentation/src/default_theme/assets/site.js' {
 declare module 'documentation/src/default_theme/assets/split.js' {
   declare module.exports: $Exports<'documentation/src/default_theme/assets/split'>;
 }
+declare module 'documentation/src/default_theme/index' {
+  declare module.exports: $Exports<'documentation/src/default_theme'>;
+}
 declare module 'documentation/src/default_theme/index.js' {
-  declare module.exports: $Exports<'documentation/src/default_theme/index'>;
+  declare module.exports: $Exports<'documentation/src/default_theme'>;
 }
 declare module 'documentation/src/extractors/comments.js' {
   declare module.exports: $Exports<'documentation/src/extractors/comments'>;
@@ -1384,6 +336,9 @@ declare module 'documentation/src/flow_doctrine.js' {
 declare module 'documentation/src/garbage_collect.js' {
   declare module.exports: $Exports<'documentation/src/garbage_collect'>;
 }
+declare module 'documentation/src/get-readme-file.js' {
+  declare module.exports: $Exports<'documentation/src/get-readme-file'>;
+}
 declare module 'documentation/src/git/find_git.js' {
   declare module.exports: $Exports<'documentation/src/git/find_git'>;
 }
@@ -1396,8 +351,11 @@ declare module 'documentation/src/github.js' {
 declare module 'documentation/src/hierarchy.js' {
   declare module.exports: $Exports<'documentation/src/hierarchy'>;
 }
+declare module 'documentation/src/index' {
+  declare module.exports: $Exports<'documentation/src'>;
+}
 declare module 'documentation/src/index.js' {
-  declare module.exports: $Exports<'documentation/src/index'>;
+  declare module.exports: $Exports<'documentation/src'>;
 }
 declare module 'documentation/src/infer/access.js' {
   declare module.exports: $Exports<'documentation/src/infer/access'>;
@@ -1407,6 +365,9 @@ declare module 'documentation/src/infer/augments.js' {
 }
 declare module 'documentation/src/infer/finders.js' {
   declare module.exports: $Exports<'documentation/src/infer/finders'>;
+}
+declare module 'documentation/src/infer/implements.js' {
+  declare module.exports: $Exports<'documentation/src/infer/implements'>;
 }
 declare module 'documentation/src/infer/kind.js' {
   declare module.exports: $Exports<'documentation/src/infer/kind'>;
@@ -1428,9 +389,6 @@ declare module 'documentation/src/infer/return.js' {
 }
 declare module 'documentation/src/infer/type.js' {
   declare module.exports: $Exports<'documentation/src/infer/type'>;
-}
-declare module 'documentation/src/inline_tokenizer.js' {
-  declare module.exports: $Exports<'documentation/src/inline_tokenizer'>;
 }
 declare module 'documentation/src/input/dependency.js' {
   declare module.exports: $Exports<'documentation/src/input/dependency'>;
@@ -1480,9 +438,6 @@ declare module 'documentation/src/output/util/linker_stack.js' {
 declare module 'documentation/src/output/util/reroute_links.js' {
   declare module.exports: $Exports<'documentation/src/output/util/reroute_links'>;
 }
-declare module 'documentation/src/parse_markdown.js' {
-  declare module.exports: $Exports<'documentation/src/parse_markdown'>;
-}
 declare module 'documentation/src/parse.js' {
   declare module.exports: $Exports<'documentation/src/parse'>;
 }
@@ -1491,6 +446,18 @@ declare module 'documentation/src/parsers/javascript.js' {
 }
 declare module 'documentation/src/parsers/parse_to_ast.js' {
   declare module.exports: $Exports<'documentation/src/parsers/parse_to_ast'>;
+}
+declare module 'documentation/src/parsers/vue.js' {
+  declare module.exports: $Exports<'documentation/src/parsers/vue'>;
+}
+declare module 'documentation/src/remark-jsDoc-link.js' {
+  declare module.exports: $Exports<'documentation/src/remark-jsDoc-link'>;
+}
+declare module 'documentation/src/remark-parse.js' {
+  declare module.exports: $Exports<'documentation/src/remark-parse'>;
+}
+declare module 'documentation/src/remark-remove-position.js' {
+  declare module.exports: $Exports<'documentation/src/remark-remove-position'>;
 }
 declare module 'documentation/src/serve/error_page.js' {
   declare module.exports: $Exports<'documentation/src/serve/error_page'>;
@@ -1503,6 +470,12 @@ declare module 'documentation/src/smart_glob.js' {
 }
 declare module 'documentation/src/sort.js' {
   declare module.exports: $Exports<'documentation/src/sort'>;
+}
+declare module 'documentation/src/ts_doctrine.js' {
+  declare module.exports: $Exports<'documentation/src/ts_doctrine'>;
+}
+declare module 'documentation/src/type_annotation.js' {
+  declare module.exports: $Exports<'documentation/src/type_annotation'>;
 }
 declare module 'documentation/src/walk.js' {
   declare module.exports: $Exports<'documentation/src/walk'>;
