@@ -1,5 +1,5 @@
-// flow-typed signature: 9a3f9c537003678c329fab4ad3bb1bd5
-// flow-typed version: <<STUB>>/eslint-config-bunchtogether_v^1.1.4/flow_v0.63.1
+// flow-typed signature: 9b38266ed84b00e58d2998418645438d
+// flow-typed version: <<STUB>>/eslint-config-bunchtogether_v^1.1.6/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
