@@ -1,5 +1,5 @@
-// flow-typed signature: 265be42ff963e1f156aad35a7f5121b4
-// flow-typed version: <<STUB>>/eslint-plugin-flowtype-errors_v^3.3.7/flow_v0.63.1
+// flow-typed signature: a33e530d4d74bf0a7ca5f66c45423855
+// flow-typed version: <<STUB>>/eslint-plugin-flowtype-errors_v^4.5.0/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -26,7 +26,7 @@ declare module 'eslint-plugin-flowtype-errors/dist/collect' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-flowtype-errors/dist/filter' {
+declare module 'eslint-plugin-flowtype-errors/dist/config/recommended' {
   declare module.exports: any;
 }
 
@@ -38,7 +38,7 @@ declare module 'eslint-plugin-flowtype-errors/dist/get-program' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-flowtype-errors/dist/index' {
+declare module 'eslint-plugin-flowtype-errors/dist' {
   declare module.exports: any;
 }
 
@@ -46,8 +46,8 @@ declare module 'eslint-plugin-flowtype-errors/dist/index' {
 declare module 'eslint-plugin-flowtype-errors/dist/collect.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist/collect'>;
 }
-declare module 'eslint-plugin-flowtype-errors/dist/filter.js' {
-  declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist/filter'>;
+declare module 'eslint-plugin-flowtype-errors/dist/config/recommended.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist/config/recommended'>;
 }
 declare module 'eslint-plugin-flowtype-errors/dist/format.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist/format'>;
@@ -55,8 +55,11 @@ declare module 'eslint-plugin-flowtype-errors/dist/format.js' {
 declare module 'eslint-plugin-flowtype-errors/dist/get-program.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist/get-program'>;
 }
+declare module 'eslint-plugin-flowtype-errors/dist/index' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist'>;
+}
 declare module 'eslint-plugin-flowtype-errors/dist/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist/index'>;
+  declare module.exports: $Exports<'eslint-plugin-flowtype-errors/dist'>;
 }
 declare module 'eslint-plugin-flowtype-errors/index' {
   declare module.exports: $Exports<'eslint-plugin-flowtype-errors'>;
