@@ -1,5 +1,5 @@
-// flow-typed signature: fecea0764b77ab496934d14cf729f960
-// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.41.0/flow_v0.63.1
+// flow-typed signature: 074b35c04629a774361274e81f376c05
+// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^8.0.3/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,11 +22,47 @@ declare module 'eslint-plugin-flowtype' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'eslint-plugin-flowtype/bin/readmeAssertions' {
+declare module 'eslint-plugin-flowtype/dist/bin/addAssertions' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-flowtype/dist/index' {
+declare module 'eslint-plugin-flowtype/dist/bin/checkDocs' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/bin/checkTests' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/bin/utilities' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle/isSimpleType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle/needWrap' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyleComplexType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyleSimpleType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/arrowParens' {
   declare module.exports: any;
 }
 
@@ -42,7 +78,19 @@ declare module 'eslint-plugin-flowtype/dist/rules/delimiterDangle' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/rules/enforceLineBreak' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/rules/genericSpacing' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/interfaceIdMatch' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/newlineAfterFlowAnnotation' {
   declare module.exports: any;
 }
 
@@ -50,7 +98,23 @@ declare module 'eslint-plugin-flowtype/dist/rules/noDupeKeys' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/rules/noDuplicateTypeUnionIntersectionMembers' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/noExistentialType' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/rules/noFlowFixMeComments' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/noInternalFlowType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/noMixed' {
   declare module.exports: any;
 }
 
@@ -74,7 +138,31 @@ declare module 'eslint-plugin-flowtype/dist/rules/noWeakTypes' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/rules/objectTypeCurlySpacing' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/rules/objectTypeDelimiter' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/quotes' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/requireCompoundTypeAlias' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/requireExactType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/requireIndexerName' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/requireInexactType' {
   declare module.exports: any;
 }
 
@@ -82,7 +170,15 @@ declare module 'eslint-plugin-flowtype/dist/rules/requireParameterType' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/rules/requireReadonlyReactProps' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/rules/requireReturnType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/requireTypesAtTop' {
   declare module.exports: any;
 }
 
@@ -102,6 +198,10 @@ declare module 'eslint-plugin-flowtype/dist/rules/sortKeys' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/rules/sortTypeUnionIntersectionMembers' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/rules/spaceAfterTypeColon' {
   declare module.exports: any;
 }
@@ -111,6 +211,10 @@ declare module 'eslint-plugin-flowtype/dist/rules/spaceBeforeGenericBracket' {
 }
 
 declare module 'eslint-plugin-flowtype/dist/rules/spaceBeforeTypeColon' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/spreadExactType' {
   declare module.exports: any;
 }
 
@@ -138,7 +242,11 @@ declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateTypic
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/index' {
+declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateVariables' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing' {
   declare module.exports: any;
 }
 
@@ -150,11 +258,19 @@ declare module 'eslint-plugin-flowtype/dist/rules/typeIdMatch' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/rules/typeImportStyle' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/rules/unionIntersectionSpacing' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-flowtype/dist/rules/useFlowType' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/rules/useReadOnlySpread' {
   declare module.exports: any;
 }
 
@@ -170,6 +286,10 @@ declare module 'eslint-plugin-flowtype/dist/utilities/fuzzyStringMatch' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-flowtype/dist/utilities/getBuiltinRule' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-flowtype/dist/utilities/getParameterName' {
   declare module.exports: any;
 }
@@ -182,7 +302,7 @@ declare module 'eslint-plugin-flowtype/dist/utilities/getTokenBeforeParens' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-flowtype/dist/utilities/index' {
+declare module 'eslint-plugin-flowtype/dist/utilities' {
   declare module.exports: any;
 }
 
@@ -191,6 +311,14 @@ declare module 'eslint-plugin-flowtype/dist/utilities/isFlowFile' {
 }
 
 declare module 'eslint-plugin-flowtype/dist/utilities/isFlowFileAnnotation' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/utilities/isNoFlowFile' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-flowtype/dist/utilities/isNoFlowFileAnnotation' {
   declare module.exports: any;
 }
 
@@ -207,11 +335,44 @@ declare module 'eslint-plugin-flowtype/dist/utilities/spacingFixers' {
 }
 
 // Filename aliases
-declare module 'eslint-plugin-flowtype/bin/readmeAssertions.js' {
-  declare module.exports: $Exports<'eslint-plugin-flowtype/bin/readmeAssertions'>;
+declare module 'eslint-plugin-flowtype/dist/bin/addAssertions.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/bin/addAssertions'>;
+}
+declare module 'eslint-plugin-flowtype/dist/bin/checkDocs.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/bin/checkDocs'>;
+}
+declare module 'eslint-plugin-flowtype/dist/bin/checkTests.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/bin/checkTests'>;
+}
+declare module 'eslint-plugin-flowtype/dist/bin/utilities.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/bin/utilities'>;
+}
+declare module 'eslint-plugin-flowtype/dist/index' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist'>;
 }
 declare module 'eslint-plugin-flowtype/dist/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/index'>;
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle/index' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrayStyle'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle/index.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrayStyle'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle/isSimpleType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrayStyle/isSimpleType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyle/needWrap.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrayStyle/needWrap'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyleComplexType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrayStyleComplexType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrayStyleSimpleType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrayStyleSimpleType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/arrowParens.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/arrowParens'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/booleanStyle.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/booleanStyle'>;
@@ -222,14 +383,35 @@ declare module 'eslint-plugin-flowtype/dist/rules/defineFlowType.js' {
 declare module 'eslint-plugin-flowtype/dist/rules/delimiterDangle.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/delimiterDangle'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/enforceLineBreak.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/enforceLineBreak'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/genericSpacing.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/genericSpacing'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/interfaceIdMatch.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/interfaceIdMatch'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/newlineAfterFlowAnnotation.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/newlineAfterFlowAnnotation'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/noDupeKeys.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noDupeKeys'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/noDuplicateTypeUnionIntersectionMembers.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noDuplicateTypeUnionIntersectionMembers'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/noExistentialType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noExistentialType'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/noFlowFixMeComments.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noFlowFixMeComments'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/noInternalFlowType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noInternalFlowType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/noMixed.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noMixed'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/noMutableArray.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noMutableArray'>;
@@ -246,14 +428,38 @@ declare module 'eslint-plugin-flowtype/dist/rules/noUnusedExpressions.js' {
 declare module 'eslint-plugin-flowtype/dist/rules/noWeakTypes.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/noWeakTypes'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/objectTypeCurlySpacing.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/objectTypeCurlySpacing'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/objectTypeDelimiter.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/objectTypeDelimiter'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/quotes.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/quotes'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/requireCompoundTypeAlias.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireCompoundTypeAlias'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/requireExactType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireExactType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/requireIndexerName.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireIndexerName'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/requireInexactType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireInexactType'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/requireParameterType.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireParameterType'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/requireReadonlyReactProps.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireReadonlyReactProps'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/requireReturnType.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireReturnType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/requireTypesAtTop.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireTypesAtTop'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/requireValidFileAnnotation.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/requireValidFileAnnotation'>;
@@ -267,6 +473,9 @@ declare module 'eslint-plugin-flowtype/dist/rules/semi.js' {
 declare module 'eslint-plugin-flowtype/dist/rules/sortKeys.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/sortKeys'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/sortTypeUnionIntersectionMembers.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/sortTypeUnionIntersectionMembers'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/spaceAfterTypeColon.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/spaceAfterTypeColon'>;
 }
@@ -275,6 +484,9 @@ declare module 'eslint-plugin-flowtype/dist/rules/spaceBeforeGenericBracket.js' 
 }
 declare module 'eslint-plugin-flowtype/dist/rules/spaceBeforeTypeColon.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/spaceBeforeTypeColon'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/spreadExactType.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/spreadExactType'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateFunctions.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateFunctions'>;
@@ -294,8 +506,14 @@ declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateTypeC
 declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateTypical.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateTypical'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateVariables.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing/evaluateVariables'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/index' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing/index'>;
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/reporter.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeColonSpacing/reporter'>;
@@ -303,11 +521,17 @@ declare module 'eslint-plugin-flowtype/dist/rules/typeColonSpacing/reporter.js' 
 declare module 'eslint-plugin-flowtype/dist/rules/typeIdMatch.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeIdMatch'>;
 }
+declare module 'eslint-plugin-flowtype/dist/rules/typeImportStyle.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/typeImportStyle'>;
+}
 declare module 'eslint-plugin-flowtype/dist/rules/unionIntersectionSpacing.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/unionIntersectionSpacing'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/useFlowType.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/useFlowType'>;
+}
+declare module 'eslint-plugin-flowtype/dist/rules/useReadOnlySpread.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/useReadOnlySpread'>;
 }
 declare module 'eslint-plugin-flowtype/dist/rules/validSyntax.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/rules/validSyntax'>;
@@ -318,6 +542,9 @@ declare module 'eslint-plugin-flowtype/dist/utilities/checkFlowFileAnnotation.js
 declare module 'eslint-plugin-flowtype/dist/utilities/fuzzyStringMatch.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/fuzzyStringMatch'>;
 }
+declare module 'eslint-plugin-flowtype/dist/utilities/getBuiltinRule.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/getBuiltinRule'>;
+}
 declare module 'eslint-plugin-flowtype/dist/utilities/getParameterName.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/getParameterName'>;
 }
@@ -327,14 +554,23 @@ declare module 'eslint-plugin-flowtype/dist/utilities/getTokenAfterParens.js' {
 declare module 'eslint-plugin-flowtype/dist/utilities/getTokenBeforeParens.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/getTokenBeforeParens'>;
 }
+declare module 'eslint-plugin-flowtype/dist/utilities/index' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities'>;
+}
 declare module 'eslint-plugin-flowtype/dist/utilities/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/index'>;
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities'>;
 }
 declare module 'eslint-plugin-flowtype/dist/utilities/isFlowFile.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/isFlowFile'>;
 }
 declare module 'eslint-plugin-flowtype/dist/utilities/isFlowFileAnnotation.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/isFlowFileAnnotation'>;
+}
+declare module 'eslint-plugin-flowtype/dist/utilities/isNoFlowFile.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/isNoFlowFile'>;
+}
+declare module 'eslint-plugin-flowtype/dist/utilities/isNoFlowFileAnnotation.js' {
+  declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/isNoFlowFileAnnotation'>;
 }
 declare module 'eslint-plugin-flowtype/dist/utilities/iterateFunctionNodes.js' {
   declare module.exports: $Exports<'eslint-plugin-flowtype/dist/utilities/iterateFunctionNodes'>;
