@@ -1,5 +1,5 @@
-// flow-typed signature: 8d6b3f06d2ca167dd08012b19ec8cabb
-// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.8.0/flow_v0.63.1
+// flow-typed signature: e7abebfda84318548986b9eccc6978c5
+// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.25.4/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -46,6 +46,10 @@ declare module 'eslint-plugin-import/config/stage-0' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/config/typescript' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/config/warnings' {
   declare module.exports: any;
 }
@@ -54,7 +58,15 @@ declare module 'eslint-plugin-import/lib/core/importType' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/lib/core/packagePath' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/lib/core/staticRequire' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/docsUrl' {
   declare module.exports: any;
 }
 
@@ -66,11 +78,15 @@ declare module 'eslint-plugin-import/lib/importDeclaration' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-import/lib/index' {
+declare module 'eslint-plugin-import/lib' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-import/lib/rules/default' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/dynamic-import-chunkname' {
   declare module.exports: any;
 }
 
@@ -87,6 +103,10 @@ declare module 'eslint-plugin-import/lib/rules/extensions' {
 }
 
 declare module 'eslint-plugin-import/lib/rules/first' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/group-exports' {
   declare module.exports: any;
 }
 
@@ -126,6 +146,14 @@ declare module 'eslint-plugin-import/lib/rules/no-commonjs' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/lib/rules/no-cycle' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-default-export' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/lib/rules/no-deprecated' {
   declare module.exports: any;
 }
@@ -139,6 +167,10 @@ declare module 'eslint-plugin-import/lib/rules/no-dynamic-require' {
 }
 
 declare module 'eslint-plugin-import/lib/rules/no-extraneous-dependencies' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-import-module-exports' {
   declare module.exports: any;
 }
 
@@ -162,6 +194,10 @@ declare module 'eslint-plugin-import/lib/rules/no-named-default' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/lib/rules/no-named-export' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/lib/rules/no-namespace' {
   declare module.exports: any;
 }
@@ -170,7 +206,19 @@ declare module 'eslint-plugin-import/lib/rules/no-nodejs-modules' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-import/lib/rules/no-relative-packages' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-relative-parent-imports' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-import/lib/rules/no-restricted-paths' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-self-import' {
   declare module.exports: any;
 }
 
@@ -179,6 +227,14 @@ declare module 'eslint-plugin-import/lib/rules/no-unassigned-import' {
 }
 
 declare module 'eslint-plugin-import/lib/rules/no-unresolved' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-unused-modules' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-import/lib/rules/no-useless-path-segments' {
   declare module.exports: any;
 }
 
@@ -198,7 +254,7 @@ declare module 'eslint-plugin-import/lib/rules/unambiguous' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-import/memo-parser/index' {
+declare module 'eslint-plugin-import/memo-parser' {
   declare module.exports: any;
 }
 
@@ -221,14 +277,23 @@ declare module 'eslint-plugin-import/config/recommended.js' {
 declare module 'eslint-plugin-import/config/stage-0.js' {
   declare module.exports: $Exports<'eslint-plugin-import/config/stage-0'>;
 }
+declare module 'eslint-plugin-import/config/typescript.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/config/typescript'>;
+}
 declare module 'eslint-plugin-import/config/warnings.js' {
   declare module.exports: $Exports<'eslint-plugin-import/config/warnings'>;
 }
 declare module 'eslint-plugin-import/lib/core/importType.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/core/importType'>;
 }
+declare module 'eslint-plugin-import/lib/core/packagePath.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/core/packagePath'>;
+}
 declare module 'eslint-plugin-import/lib/core/staticRequire.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/core/staticRequire'>;
+}
+declare module 'eslint-plugin-import/lib/docsUrl.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/docsUrl'>;
 }
 declare module 'eslint-plugin-import/lib/ExportMap.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/ExportMap'>;
@@ -236,11 +301,17 @@ declare module 'eslint-plugin-import/lib/ExportMap.js' {
 declare module 'eslint-plugin-import/lib/importDeclaration.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/importDeclaration'>;
 }
+declare module 'eslint-plugin-import/lib/index' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib'>;
+}
 declare module 'eslint-plugin-import/lib/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/index'>;
+  declare module.exports: $Exports<'eslint-plugin-import/lib'>;
 }
 declare module 'eslint-plugin-import/lib/rules/default.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/default'>;
+}
+declare module 'eslint-plugin-import/lib/rules/dynamic-import-chunkname.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/dynamic-import-chunkname'>;
 }
 declare module 'eslint-plugin-import/lib/rules/export.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/export'>;
@@ -253,6 +324,9 @@ declare module 'eslint-plugin-import/lib/rules/extensions.js' {
 }
 declare module 'eslint-plugin-import/lib/rules/first.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/first'>;
+}
+declare module 'eslint-plugin-import/lib/rules/group-exports.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/group-exports'>;
 }
 declare module 'eslint-plugin-import/lib/rules/imports-first.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/imports-first'>;
@@ -281,6 +355,12 @@ declare module 'eslint-plugin-import/lib/rules/no-anonymous-default-export.js' {
 declare module 'eslint-plugin-import/lib/rules/no-commonjs.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-commonjs'>;
 }
+declare module 'eslint-plugin-import/lib/rules/no-cycle.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-cycle'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-default-export.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-default-export'>;
+}
 declare module 'eslint-plugin-import/lib/rules/no-deprecated.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-deprecated'>;
 }
@@ -292,6 +372,9 @@ declare module 'eslint-plugin-import/lib/rules/no-dynamic-require.js' {
 }
 declare module 'eslint-plugin-import/lib/rules/no-extraneous-dependencies.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-extraneous-dependencies'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-import-module-exports.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-import-module-exports'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-internal-modules.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-internal-modules'>;
@@ -308,20 +391,38 @@ declare module 'eslint-plugin-import/lib/rules/no-named-as-default.js' {
 declare module 'eslint-plugin-import/lib/rules/no-named-default.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-named-default'>;
 }
+declare module 'eslint-plugin-import/lib/rules/no-named-export.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-named-export'>;
+}
 declare module 'eslint-plugin-import/lib/rules/no-namespace.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-namespace'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-nodejs-modules.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-nodejs-modules'>;
 }
+declare module 'eslint-plugin-import/lib/rules/no-relative-packages.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-relative-packages'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-relative-parent-imports.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-relative-parent-imports'>;
+}
 declare module 'eslint-plugin-import/lib/rules/no-restricted-paths.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-restricted-paths'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-self-import.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-self-import'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-unassigned-import.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unassigned-import'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-unresolved.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unresolved'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-unused-modules.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unused-modules'>;
+}
+declare module 'eslint-plugin-import/lib/rules/no-useless-path-segments.js' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-useless-path-segments'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-webpack-loader-syntax.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-webpack-loader-syntax'>;
@@ -335,6 +436,9 @@ declare module 'eslint-plugin-import/lib/rules/prefer-default-export.js' {
 declare module 'eslint-plugin-import/lib/rules/unambiguous.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/unambiguous'>;
 }
+declare module 'eslint-plugin-import/memo-parser/index' {
+  declare module.exports: $Exports<'eslint-plugin-import/memo-parser'>;
+}
 declare module 'eslint-plugin-import/memo-parser/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/memo-parser/index'>;
+  declare module.exports: $Exports<'eslint-plugin-import/memo-parser'>;
 }
