@@ -1,5 +1,5 @@
-// flow-typed signature: 2f43267d497688ac76b5143225c08c82
-// flow-typed version: <<STUB>>/eslint-plugin-jest_v^21.7.0/flow_v0.63.1
+// flow-typed signature: df933ec6353210ca65235ce77d4af029
+// flow-typed version: <<STUB>>/eslint-plugin-jest_v^26.0.0/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,171 +22,350 @@ declare module 'eslint-plugin-jest' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'eslint-plugin-jest/processors/__tests__/snapshot-processor.test' {
+declare module 'eslint-plugin-jest/lib' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/processors/snapshot-processor' {
+declare module 'eslint-plugin-jest/lib/processors/snapshot-processor' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/no_focused_tests.test' {
+declare module 'eslint-plugin-jest/lib/rules/consistent-test-it' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/no_identical_title.test' {
+declare module 'eslint-plugin-jest/lib/rules/detectJestVersion' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/no_large_snapshots.test' {
+declare module 'eslint-plugin-jest/lib/rules/expect-expect' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/no_skipped_tests.test' {
+declare module 'eslint-plugin-jest/lib/rules/max-nested-describe' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_expect_assertions.test' {
+declare module 'eslint-plugin-jest/lib/rules/no-alias-methods' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_be_null.test' {
+declare module 'eslint-plugin-jest/lib/rules/no-commented-out-tests' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_be_undefined.test' {
+declare module 'eslint-plugin-jest/lib/rules/no-conditional-expect' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_have_length.test' {
+declare module 'eslint-plugin-jest/lib/rules/no-deprecated-functions' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/valid_expect_in_promise.test' {
+declare module 'eslint-plugin-jest/lib/rules/no-disabled-tests' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/__tests__/valid_expect.test' {
+declare module 'eslint-plugin-jest/lib/rules/no-done-callback' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_disabled_tests' {
+declare module 'eslint-plugin-jest/lib/rules/no-duplicate-hooks' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_focused_tests' {
+declare module 'eslint-plugin-jest/lib/rules/no-export' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_identical_title' {
+declare module 'eslint-plugin-jest/lib/rules/no-focused-tests' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/no_large_snapshots' {
+declare module 'eslint-plugin-jest/lib/rules/no-hooks' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/prefer_expect_assertions' {
+declare module 'eslint-plugin-jest/lib/rules/no-identical-title' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/prefer_to_be_null' {
+declare module 'eslint-plugin-jest/lib/rules/no-if' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/prefer_to_be_undefined' {
+declare module 'eslint-plugin-jest/lib/rules/no-interpolation-in-snapshots' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/prefer_to_have_length' {
+declare module 'eslint-plugin-jest/lib/rules/no-jasmine-globals' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/util' {
+declare module 'eslint-plugin-jest/lib/rules/no-jest-import' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/valid_expect_in_promise' {
+declare module 'eslint-plugin-jest/lib/rules/no-large-snapshots' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jest/rules/valid_expect' {
+declare module 'eslint-plugin-jest/lib/rules/no-mocks-import' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/no-restricted-matchers' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/no-standalone-expect' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/no-test-prefixes' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/no-test-return-statement' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-called-with' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-comparison-matcher' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-equality-matcher' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-expect-assertions' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-expect-resolves' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-hooks-on-top' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-lowercase-title' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-spy-on' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-strict-equal' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-to-be' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-to-contain' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-to-have-length' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/prefer-todo' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/require-hook' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/require-to-throw-message' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/require-top-level-describe' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/unbound-method' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/utils' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/valid-describe-callback' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/valid-expect-in-promise' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/valid-expect' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jest/lib/rules/valid-title' {
   declare module.exports: any;
 }
 
 // Filename aliases
-declare module 'eslint-plugin-jest/index' {
-  declare module.exports: $Exports<'eslint-plugin-jest'>;
+declare module 'eslint-plugin-jest/lib/index' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib'>;
 }
-declare module 'eslint-plugin-jest/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest'>;
+declare module 'eslint-plugin-jest/lib/index.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib'>;
 }
-declare module 'eslint-plugin-jest/processors/__tests__/snapshot-processor.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/processors/__tests__/snapshot-processor.test'>;
+declare module 'eslint-plugin-jest/lib/processors/snapshot-processor.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/processors/snapshot-processor'>;
 }
-declare module 'eslint-plugin-jest/processors/snapshot-processor.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/processors/snapshot-processor'>;
+declare module 'eslint-plugin-jest/lib/rules/consistent-test-it.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/consistent-test-it'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_focused_tests.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_focused_tests.test'>;
+declare module 'eslint-plugin-jest/lib/rules/detectJestVersion.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/detectJestVersion'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_identical_title.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_identical_title.test'>;
+declare module 'eslint-plugin-jest/lib/rules/expect-expect.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/expect-expect'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_large_snapshots.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_large_snapshots.test'>;
+declare module 'eslint-plugin-jest/lib/rules/max-nested-describe.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/max-nested-describe'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/no_skipped_tests.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/no_skipped_tests.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-alias-methods.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-alias-methods'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_expect_assertions.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer_expect_assertions.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-commented-out-tests.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-commented-out-tests'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_be_null.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer_to_be_null.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-conditional-expect.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-conditional-expect'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_be_undefined.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer_to_be_undefined.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-deprecated-functions.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-deprecated-functions'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/prefer_to_have_length.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/prefer_to_have_length.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-disabled-tests.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-disabled-tests'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/valid_expect_in_promise.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/valid_expect_in_promise.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-done-callback.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-done-callback'>;
 }
-declare module 'eslint-plugin-jest/rules/__tests__/valid_expect.test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/__tests__/valid_expect.test'>;
+declare module 'eslint-plugin-jest/lib/rules/no-duplicate-hooks.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-duplicate-hooks'>;
 }
-declare module 'eslint-plugin-jest/rules/no_disabled_tests.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_disabled_tests'>;
+declare module 'eslint-plugin-jest/lib/rules/no-export.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-export'>;
 }
-declare module 'eslint-plugin-jest/rules/no_focused_tests.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_focused_tests'>;
+declare module 'eslint-plugin-jest/lib/rules/no-focused-tests.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-focused-tests'>;
 }
-declare module 'eslint-plugin-jest/rules/no_identical_title.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_identical_title'>;
+declare module 'eslint-plugin-jest/lib/rules/no-hooks.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-hooks'>;
 }
-declare module 'eslint-plugin-jest/rules/no_large_snapshots.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/no_large_snapshots'>;
+declare module 'eslint-plugin-jest/lib/rules/no-identical-title.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-identical-title'>;
 }
-declare module 'eslint-plugin-jest/rules/prefer_expect_assertions.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer_expect_assertions'>;
+declare module 'eslint-plugin-jest/lib/rules/no-if.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-if'>;
 }
-declare module 'eslint-plugin-jest/rules/prefer_to_be_null.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer_to_be_null'>;
+declare module 'eslint-plugin-jest/lib/rules/no-interpolation-in-snapshots.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-interpolation-in-snapshots'>;
 }
-declare module 'eslint-plugin-jest/rules/prefer_to_be_undefined.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer_to_be_undefined'>;
+declare module 'eslint-plugin-jest/lib/rules/no-jasmine-globals.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-jasmine-globals'>;
 }
-declare module 'eslint-plugin-jest/rules/prefer_to_have_length.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/prefer_to_have_length'>;
+declare module 'eslint-plugin-jest/lib/rules/no-jest-import.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-jest-import'>;
 }
-declare module 'eslint-plugin-jest/rules/util.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/util'>;
+declare module 'eslint-plugin-jest/lib/rules/no-large-snapshots.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-large-snapshots'>;
 }
-declare module 'eslint-plugin-jest/rules/valid_expect_in_promise.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/valid_expect_in_promise'>;
+declare module 'eslint-plugin-jest/lib/rules/no-mocks-import.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-mocks-import'>;
 }
-declare module 'eslint-plugin-jest/rules/valid_expect.js' {
-  declare module.exports: $Exports<'eslint-plugin-jest/rules/valid_expect'>;
+declare module 'eslint-plugin-jest/lib/rules/no-restricted-matchers.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-restricted-matchers'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/no-standalone-expect.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-standalone-expect'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/no-test-prefixes.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-test-prefixes'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/no-test-return-statement.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/no-test-return-statement'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-called-with.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-called-with'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-comparison-matcher.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-comparison-matcher'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-equality-matcher.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-equality-matcher'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-expect-assertions.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-expect-assertions'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-expect-resolves.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-expect-resolves'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-hooks-on-top.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-hooks-on-top'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-lowercase-title.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-lowercase-title'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-spy-on.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-spy-on'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-strict-equal.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-strict-equal'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-to-be.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-to-be'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-to-contain.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-to-contain'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-to-have-length.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-to-have-length'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/prefer-todo.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/prefer-todo'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/require-hook.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/require-hook'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/require-to-throw-message.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/require-to-throw-message'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/require-top-level-describe.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/require-top-level-describe'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/unbound-method.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/unbound-method'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/utils.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/utils'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/valid-describe-callback.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/valid-describe-callback'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/valid-expect-in-promise.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/valid-expect-in-promise'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/valid-expect.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/valid-expect'>;
+}
+declare module 'eslint-plugin-jest/lib/rules/valid-title.js' {
+  declare module.exports: $Exports<'eslint-plugin-jest/lib/rules/valid-title'>;
 }
