@@ -1,5 +1,5 @@
-// flow-typed signature: 4004db3bfc4c0a4f9543a1454d608add
-// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v^6.0.3/flow_v0.63.1
+// flow-typed signature: 5a830e8db32c8d2cb953097db66dfd32
+// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v^6.5.1/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -39,6 +39,22 @@ declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock' {
 }
 
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXSpreadAttributeMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXTextMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/LiteralMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/axeMapping' {
   declare module.exports: any;
 }
 
@@ -90,7 +106,15 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elem
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/autocomplete-valid-test' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/control-has-associated-label-test' {
   declare module.exports: any;
 }
 
@@ -111,6 +135,10 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-tes
 }
 
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/interactive-supports-focus-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-associated-control-test' {
   declare module.exports: any;
 }
 
@@ -190,6 +218,18 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/attributesComparator-t
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getComputedRole-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getExplicitRole-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getImplicitRole-test' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test' {
   declare module.exports: any;
 }
@@ -218,6 +258,14 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isDisabledElement-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isDOMElement-test' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test' {
   declare module.exports: any;
 }
@@ -234,6 +282,22 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-t
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonLiteralProperty-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/mayContainChildComponent-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/mayHaveAccessibleLabel-test' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test' {
   declare module.exports: any;
 }
@@ -242,7 +306,7 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/lib/index' {
+declare module 'eslint-plugin-jsx-a11y/lib' {
   declare module.exports: any;
 }
 
@@ -282,7 +346,15 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/rules/autocomplete-valid' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/rules/click-events-have-key-events' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/rules/control-has-associated-label' {
   declare module.exports: any;
 }
 
@@ -303,6 +375,10 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/img-redundant-alt' {
 }
 
 declare module 'eslint-plugin-jsx-a11y/lib/rules/interactive-supports-focus' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/rules/label-has-associated-control' {
   declare module.exports: any;
 }
 
@@ -379,6 +455,14 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/tabindex-no-positive' {
 }
 
 declare module 'eslint-plugin-jsx-a11y/lib/util/attributesComparator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/getComputedRole' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/getExplicitRole' {
   declare module.exports: any;
 }
 
@@ -474,7 +558,7 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index' {
+declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles' {
   declare module.exports: any;
 }
 
@@ -554,6 +638,14 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/isAbstractRole' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/util/isDisabledElement' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/isDOMElement' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader' {
   declare module.exports: any;
 }
@@ -574,7 +666,23 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/util/isNonLiteralProperty' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/util/isPresentationRole' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/isSemanticRoleElement' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/mayContainChildComponent' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/mayHaveAccessibleLabel' {
   declare module.exports: any;
 }
 
@@ -618,6 +726,18 @@ declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock.js' {
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock'>;
 }
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXSpreadAttributeMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXSpreadAttributeMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXTextMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXTextMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__mocks__/LiteralMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/LiteralMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/axeMapping.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/__util__/axeMapping'>;
+}
 declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper'>;
 }
@@ -654,8 +774,14 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test.js' {
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test'>;
 }
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/autocomplete-valid-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/autocomplete-valid-test'>;
+}
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/control-has-associated-label-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/control-has-associated-label-test'>;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test'>;
@@ -671,6 +797,9 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-tes
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/interactive-supports-focus-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/interactive-supports-focus-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-associated-control-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-associated-control-test'>;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test'>;
@@ -729,6 +858,15 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/attributesComparator-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/attributesComparator-test'>;
 }
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getComputedRole-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getComputedRole-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getExplicitRole-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getExplicitRole-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getImplicitRole-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getImplicitRole-test'>;
+}
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test'>;
 }
@@ -750,6 +888,12 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test'>;
 }
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isDisabledElement-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isDisabledElement-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isDOMElement-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isDOMElement-test'>;
+}
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test'>;
 }
@@ -762,14 +906,29 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveElemen
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-test'>;
 }
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonLiteralProperty-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isNonLiteralProperty-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/mayContainChildComponent-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/mayContainChildComponent-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/mayHaveAccessibleLabel-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/mayHaveAccessibleLabel-test'>;
+}
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test'>;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/index' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/index'>;
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji'>;
@@ -798,8 +957,14 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-role.js' {
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/rules/autocomplete-valid.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/autocomplete-valid'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/rules/click-events-have-key-events.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/click-events-have-key-events'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/rules/control-has-associated-label.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/control-has-associated-label'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/heading-has-content.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/heading-has-content'>;
@@ -815,6 +980,9 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/img-redundant-alt.js' {
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/interactive-supports-focus.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/interactive-supports-focus'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/rules/label-has-associated-control.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/label-has-associated-control'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/label-has-for.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/label-has-for'>;
@@ -872,6 +1040,12 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/tabindex-no-positive.js' {
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/attributesComparator.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/attributesComparator'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/getComputedRole.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getComputedRole'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/getExplicitRole.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getExplicitRole'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/getImplicitRole.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getImplicitRole'>;
@@ -942,8 +1116,11 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/hr.js' {
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index'>;
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/input.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/input'>;
@@ -1002,6 +1179,12 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ul.js' {
 declare module 'eslint-plugin-jsx-a11y/lib/util/isAbstractRole.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isAbstractRole'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/util/isDisabledElement.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isDisabledElement'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/isDOMElement.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isDOMElement'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader'>;
 }
@@ -1017,8 +1200,20 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveElement.js' {
 declare module 'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/util/isNonLiteralProperty.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isNonLiteralProperty'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/util/isPresentationRole.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isPresentationRole'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/isSemanticRoleElement.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isSemanticRoleElement'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/mayContainChildComponent.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/mayContainChildComponent'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/mayHaveAccessibleLabel.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/mayHaveAccessibleLabel'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/schemas.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/schemas'>;
