@@ -1,5 +1,5 @@
-// flow-typed signature: 1051f82a8632507bc1b4d98a566be86d
-// flow-typed version: <<STUB>>/eslint-plugin-react_v^7.5.1/flow_v0.63.1
+// flow-typed signature: df1be0778611a456a4968f561aeed92d
+// flow-typed version: <<STUB>>/eslint-plugin-react_v^7.28.0/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -46,6 +46,10 @@ declare module 'eslint-plugin-react/lib/rules/forbid-component-props' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/forbid-dom-props' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/forbid-elements' {
   declare module.exports: any;
 }
@@ -58,7 +62,15 @@ declare module 'eslint-plugin-react/lib/rules/forbid-prop-types' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/function-component-definition' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/jsx-boolean-value' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-child-element-spacing' {
   declare module.exports: any;
 }
 
@@ -71,6 +83,10 @@ declare module 'eslint-plugin-react/lib/rules/jsx-closing-tag-location' {
 }
 
 declare module 'eslint-plugin-react/lib/rules/jsx-curly-brace-presence' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-curly-newline' {
   declare module.exports: any;
 }
 
@@ -90,6 +106,10 @@ declare module 'eslint-plugin-react/lib/rules/jsx-first-prop-new-line' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/jsx-fragments' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/jsx-handler-names' {
   declare module.exports: any;
 }
@@ -106,7 +126,15 @@ declare module 'eslint-plugin-react/lib/rules/jsx-key' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/jsx-max-depth' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/jsx-max-props-per-line' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-newline' {
   declare module.exports: any;
 }
 
@@ -118,11 +146,19 @@ declare module 'eslint-plugin-react/lib/rules/jsx-no-comment-textnodes' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/jsx-no-constructed-context-values' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/jsx-no-duplicate-props' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-react/lib/rules/jsx-no-literals' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-no-script-url' {
   declare module.exports: any;
 }
 
@@ -134,11 +170,27 @@ declare module 'eslint-plugin-react/lib/rules/jsx-no-undef' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/jsx-no-useless-fragment' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/jsx-one-expression-per-line' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-react/lib/rules/jsx-pascal-case' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-props-no-multi-spaces' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-props-no-spreading' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-sort-default-props' {
   declare module.exports: any;
 }
 
@@ -170,7 +222,15 @@ declare module 'eslint-plugin-react/lib/rules/no-access-state-in-setstate' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/no-adjacent-inline-elements' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/no-array-index-key' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/no-arrow-function-lifecycle' {
   declare module.exports: any;
 }
 
@@ -206,11 +266,19 @@ declare module 'eslint-plugin-react/lib/rules/no-find-dom-node' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/no-invalid-html-attribute' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/no-is-mounted' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-react/lib/rules/no-multi-comp' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/no-namespace' {
   declare module.exports: any;
 }
 
@@ -230,6 +298,10 @@ declare module 'eslint-plugin-react/lib/rules/no-string-refs' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/no-this-in-sfc' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/no-typos' {
   declare module.exports: any;
 }
@@ -239,6 +311,18 @@ declare module 'eslint-plugin-react/lib/rules/no-unescaped-entities' {
 }
 
 declare module 'eslint-plugin-react/lib/rules/no-unknown-property' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/no-unsafe' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/no-unstable-nested-components' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/no-unused-class-component-methods' {
   declare module.exports: any;
 }
 
@@ -255,6 +339,14 @@ declare module 'eslint-plugin-react/lib/rules/no-will-update-set-state' {
 }
 
 declare module 'eslint-plugin-react/lib/rules/prefer-es6-class' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/prefer-exact-props' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/prefer-read-only-props' {
   declare module.exports: any;
 }
 
@@ -294,6 +386,14 @@ declare module 'eslint-plugin-react/lib/rules/sort-prop-types' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/state-in-constructor' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/static-property-placement' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/style-prop-object' {
   declare module.exports: any;
 }
@@ -314,7 +414,47 @@ declare module 'eslint-plugin-react/lib/util/Components' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/util/defaultProps' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/docsUrl' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/error' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/isCreateElement' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/isDestructuredFromPragmaImport' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/isFirstLetterCapitalized' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/jsx' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/lifecycleMethods' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/linkComponents' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/log' {
   declare module.exports: any;
 }
 
@@ -327,6 +467,26 @@ declare module 'eslint-plugin-react/lib/util/pragma' {
 }
 
 declare module 'eslint-plugin-react/lib/util/props' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/propTypes' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/propTypesSort' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/propWrapper' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/report' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/util/usedPropTypes' {
   declare module.exports: any;
 }
 
@@ -363,6 +523,9 @@ declare module 'eslint-plugin-react/lib/rules/display-name.js' {
 declare module 'eslint-plugin-react/lib/rules/forbid-component-props.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-component-props'>;
 }
+declare module 'eslint-plugin-react/lib/rules/forbid-dom-props.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-dom-props'>;
+}
 declare module 'eslint-plugin-react/lib/rules/forbid-elements.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-elements'>;
 }
@@ -372,8 +535,14 @@ declare module 'eslint-plugin-react/lib/rules/forbid-foreign-prop-types.js' {
 declare module 'eslint-plugin-react/lib/rules/forbid-prop-types.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-prop-types'>;
 }
+declare module 'eslint-plugin-react/lib/rules/function-component-definition.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/function-component-definition'>;
+}
 declare module 'eslint-plugin-react/lib/rules/jsx-boolean-value.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-boolean-value'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-child-element-spacing.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-child-element-spacing'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-closing-bracket-location.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-closing-bracket-location'>;
@@ -383,6 +552,9 @@ declare module 'eslint-plugin-react/lib/rules/jsx-closing-tag-location.js' {
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-curly-brace-presence.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-curly-brace-presence'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-curly-newline.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-curly-newline'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-curly-spacing.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-curly-spacing'>;
@@ -396,6 +568,9 @@ declare module 'eslint-plugin-react/lib/rules/jsx-filename-extension.js' {
 declare module 'eslint-plugin-react/lib/rules/jsx-first-prop-new-line.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-first-prop-new-line'>;
 }
+declare module 'eslint-plugin-react/lib/rules/jsx-fragments.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-fragments'>;
+}
 declare module 'eslint-plugin-react/lib/rules/jsx-handler-names.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-handler-names'>;
 }
@@ -408,8 +583,14 @@ declare module 'eslint-plugin-react/lib/rules/jsx-indent.js' {
 declare module 'eslint-plugin-react/lib/rules/jsx-key.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-key'>;
 }
+declare module 'eslint-plugin-react/lib/rules/jsx-max-depth.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-max-depth'>;
+}
 declare module 'eslint-plugin-react/lib/rules/jsx-max-props-per-line.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-max-props-per-line'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-newline.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-newline'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-bind.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-bind'>;
@@ -417,11 +598,17 @@ declare module 'eslint-plugin-react/lib/rules/jsx-no-bind.js' {
 declare module 'eslint-plugin-react/lib/rules/jsx-no-comment-textnodes.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-comment-textnodes'>;
 }
+declare module 'eslint-plugin-react/lib/rules/jsx-no-constructed-context-values.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-constructed-context-values'>;
+}
 declare module 'eslint-plugin-react/lib/rules/jsx-no-duplicate-props.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-duplicate-props'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-literals.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-literals'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-no-script-url.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-script-url'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-target-blank.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-target-blank'>;
@@ -429,11 +616,23 @@ declare module 'eslint-plugin-react/lib/rules/jsx-no-target-blank.js' {
 declare module 'eslint-plugin-react/lib/rules/jsx-no-undef.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-undef'>;
 }
+declare module 'eslint-plugin-react/lib/rules/jsx-no-useless-fragment.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-useless-fragment'>;
+}
 declare module 'eslint-plugin-react/lib/rules/jsx-one-expression-per-line.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-one-expression-per-line'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-pascal-case.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-pascal-case'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-props-no-multi-spaces.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-props-no-multi-spaces'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-props-no-spreading.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-props-no-spreading'>;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-sort-default-props.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-sort-default-props'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-sort-props.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-sort-props'>;
@@ -456,8 +655,14 @@ declare module 'eslint-plugin-react/lib/rules/jsx-wrap-multilines.js' {
 declare module 'eslint-plugin-react/lib/rules/no-access-state-in-setstate.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-access-state-in-setstate'>;
 }
+declare module 'eslint-plugin-react/lib/rules/no-adjacent-inline-elements.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-adjacent-inline-elements'>;
+}
 declare module 'eslint-plugin-react/lib/rules/no-array-index-key.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-array-index-key'>;
+}
+declare module 'eslint-plugin-react/lib/rules/no-arrow-function-lifecycle.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-arrow-function-lifecycle'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-children-prop.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-children-prop'>;
@@ -483,11 +688,17 @@ declare module 'eslint-plugin-react/lib/rules/no-direct-mutation-state.js' {
 declare module 'eslint-plugin-react/lib/rules/no-find-dom-node.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-find-dom-node'>;
 }
+declare module 'eslint-plugin-react/lib/rules/no-invalid-html-attribute.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-invalid-html-attribute'>;
+}
 declare module 'eslint-plugin-react/lib/rules/no-is-mounted.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-is-mounted'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-multi-comp.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-multi-comp'>;
+}
+declare module 'eslint-plugin-react/lib/rules/no-namespace.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-namespace'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-redundant-should-component-update.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-redundant-should-component-update'>;
@@ -501,6 +712,9 @@ declare module 'eslint-plugin-react/lib/rules/no-set-state.js' {
 declare module 'eslint-plugin-react/lib/rules/no-string-refs.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-string-refs'>;
 }
+declare module 'eslint-plugin-react/lib/rules/no-this-in-sfc.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-this-in-sfc'>;
+}
 declare module 'eslint-plugin-react/lib/rules/no-typos.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-typos'>;
 }
@@ -509,6 +723,15 @@ declare module 'eslint-plugin-react/lib/rules/no-unescaped-entities.js' {
 }
 declare module 'eslint-plugin-react/lib/rules/no-unknown-property.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unknown-property'>;
+}
+declare module 'eslint-plugin-react/lib/rules/no-unsafe.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unsafe'>;
+}
+declare module 'eslint-plugin-react/lib/rules/no-unstable-nested-components.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unstable-nested-components'>;
+}
+declare module 'eslint-plugin-react/lib/rules/no-unused-class-component-methods.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unused-class-component-methods'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-unused-prop-types.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unused-prop-types'>;
@@ -521,6 +744,12 @@ declare module 'eslint-plugin-react/lib/rules/no-will-update-set-state.js' {
 }
 declare module 'eslint-plugin-react/lib/rules/prefer-es6-class.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-es6-class'>;
+}
+declare module 'eslint-plugin-react/lib/rules/prefer-exact-props.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-exact-props'>;
+}
+declare module 'eslint-plugin-react/lib/rules/prefer-read-only-props.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-read-only-props'>;
 }
 declare module 'eslint-plugin-react/lib/rules/prefer-stateless-function.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-stateless-function'>;
@@ -549,6 +778,12 @@ declare module 'eslint-plugin-react/lib/rules/sort-comp.js' {
 declare module 'eslint-plugin-react/lib/rules/sort-prop-types.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/sort-prop-types'>;
 }
+declare module 'eslint-plugin-react/lib/rules/state-in-constructor.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/state-in-constructor'>;
+}
+declare module 'eslint-plugin-react/lib/rules/static-property-placement.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/static-property-placement'>;
+}
 declare module 'eslint-plugin-react/lib/rules/style-prop-object.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/style-prop-object'>;
 }
@@ -564,8 +799,38 @@ declare module 'eslint-plugin-react/lib/util/ast.js' {
 declare module 'eslint-plugin-react/lib/util/Components.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/Components'>;
 }
+declare module 'eslint-plugin-react/lib/util/defaultProps.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/defaultProps'>;
+}
+declare module 'eslint-plugin-react/lib/util/docsUrl.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/docsUrl'>;
+}
+declare module 'eslint-plugin-react/lib/util/error.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/error'>;
+}
 declare module 'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket'>;
+}
+declare module 'eslint-plugin-react/lib/util/isCreateElement.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/isCreateElement'>;
+}
+declare module 'eslint-plugin-react/lib/util/isDestructuredFromPragmaImport.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/isDestructuredFromPragmaImport'>;
+}
+declare module 'eslint-plugin-react/lib/util/isFirstLetterCapitalized.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/isFirstLetterCapitalized'>;
+}
+declare module 'eslint-plugin-react/lib/util/jsx.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/jsx'>;
+}
+declare module 'eslint-plugin-react/lib/util/lifecycleMethods.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/lifecycleMethods'>;
+}
+declare module 'eslint-plugin-react/lib/util/linkComponents.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/linkComponents'>;
+}
+declare module 'eslint-plugin-react/lib/util/log.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/log'>;
 }
 declare module 'eslint-plugin-react/lib/util/makeNoMethodSetStateRule.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/makeNoMethodSetStateRule'>;
@@ -575,6 +840,21 @@ declare module 'eslint-plugin-react/lib/util/pragma.js' {
 }
 declare module 'eslint-plugin-react/lib/util/props.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/props'>;
+}
+declare module 'eslint-plugin-react/lib/util/propTypes.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/propTypes'>;
+}
+declare module 'eslint-plugin-react/lib/util/propTypesSort.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/propTypesSort'>;
+}
+declare module 'eslint-plugin-react/lib/util/propWrapper.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/propWrapper'>;
+}
+declare module 'eslint-plugin-react/lib/util/report.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/report'>;
+}
+declare module 'eslint-plugin-react/lib/util/usedPropTypes.js' {
+  declare module.exports: $Exports<'eslint-plugin-react/lib/util/usedPropTypes'>;
 }
 declare module 'eslint-plugin-react/lib/util/variable.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/variable'>;
