@@ -1,5 +1,5 @@
-// flow-typed signature: b422982952164f2d7a633810b77f5b44
-// flow-typed version: <<STUB>>/eslint_v^4.16.0/flow_v0.63.1
+// flow-typed signature: 55f178ac0b0f655ac6ff689e0c869512
+// flow-typed version: <<STUB>>/eslint_v^8.8.0/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -34,14 +34,6 @@ declare module 'eslint/conf/default-cli-options' {
   declare module.exports: any;
 }
 
-declare module 'eslint/conf/default-config-options' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/conf/environments' {
-  declare module.exports: any;
-}
-
 declare module 'eslint/conf/eslint-all' {
   declare module.exports: any;
 }
@@ -50,11 +42,67 @@ declare module 'eslint/conf/eslint-recommended' {
   declare module.exports: any;
 }
 
+declare module 'eslint/conf/globals' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/api' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/ast-utils' {
+declare module 'eslint/lib/cli-engine/cli-engine' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/file-enumerator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/checkstyle' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/compact' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/html' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/jslint-xml' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/json-with-metadata' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/json' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/junit' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/stylish' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/tap' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/unix' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/visualstudio' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/hash' {
   declare module.exports: any;
 }
 
@@ -62,131 +110,83 @@ declare module 'eslint/lib/cli-engine' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/cli-engine/lint-result-cache' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/load-rules' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/xml-escape' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/cli' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path-analyzer' {
+declare module 'eslint/lib/config/default-config' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path-segment' {
+declare module 'eslint/lib/config/flat-config-array' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path-state' {
+declare module 'eslint/lib/config/flat-config-helpers' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path' {
+declare module 'eslint/lib/config/flat-config-schema' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/debug-helpers' {
+declare module 'eslint/lib/config/rule-validator' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/fork-context' {
+declare module 'eslint/lib/eslint/eslint' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/id-generator' {
+declare module 'eslint/lib/eslint' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config' {
+declare module 'eslint/lib/linter/apply-disable-directives' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/autoconfig' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path-analyzer' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-cache' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path-segment' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-file' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path-state' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-initializer' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-ops' {
+declare module 'eslint/lib/linter/code-path-analysis/debug-helpers' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-rule' {
+declare module 'eslint/lib/linter/code-path-analysis/fork-context' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-validator' {
+declare module 'eslint/lib/linter/code-path-analysis/id-generator' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/environments' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/config/plugins' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/file-finder' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/checkstyle' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/codeframe' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/compact' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/html' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/jslint-xml' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/json' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/junit' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/stylish' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/table' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/tap' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/unix' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/visualstudio' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/ignored-paths' {
+declare module 'eslint/lib/linter/config-comment-parser' {
   declare module.exports: any;
 }
 
@@ -194,11 +194,39 @@ declare module 'eslint/lib/linter' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/load-rules' {
+declare module 'eslint/lib/linter/interpolate' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/logging' {
+declare module 'eslint/lib/linter/linter' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/node-event-generator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/report-translator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/rule-fixer' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/rules' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/safe-emitter' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/source-code-fixer' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/timing' {
   declare module.exports: any;
 }
 
@@ -206,11 +234,11 @@ declare module 'eslint/lib/options' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/report-translator' {
+declare module 'eslint/lib/rule-tester' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/rules' {
+declare module 'eslint/lib/rule-tester/rule-tester' {
   declare module.exports: any;
 }
 
@@ -310,7 +338,15 @@ declare module 'eslint/lib/rules/curly' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/default-case-last' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/default-case' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/default-param-last' {
   declare module.exports: any;
 }
 
@@ -350,6 +386,10 @@ declare module 'eslint/lib/rules/func-style' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/function-call-argument-newline' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/function-paren-newline' {
   declare module.exports: any;
 }
@@ -366,6 +406,10 @@ declare module 'eslint/lib/rules/global-require' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/grouped-accessor-pairs' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/guard-for-in' {
   declare module.exports: any;
 }
@@ -375,6 +419,10 @@ declare module 'eslint/lib/rules/handle-callback-err' {
 }
 
 declare module 'eslint/lib/rules/id-blacklist' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/id-denylist' {
   declare module.exports: any;
 }
 
@@ -395,6 +443,10 @@ declare module 'eslint/lib/rules/indent-legacy' {
 }
 
 declare module 'eslint/lib/rules/indent' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules' {
   declare module.exports: any;
 }
 
@@ -434,11 +486,19 @@ declare module 'eslint/lib/rules/lines-between-class-members' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/max-classes-per-file' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/max-depth' {
   declare module.exports: any;
 }
 
 declare module 'eslint/lib/rules/max-len' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/max-lines-per-function' {
   declare module.exports: any;
 }
 
@@ -498,6 +558,10 @@ declare module 'eslint/lib/rules/no-array-constructor' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-async-promise-executor' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-await-in-loop' {
   declare module.exports: any;
 }
@@ -550,6 +614,10 @@ declare module 'eslint/lib/rules/no-constant-condition' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-constructor-return' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-continue' {
   declare module.exports: any;
 }
@@ -575,6 +643,10 @@ declare module 'eslint/lib/rules/no-dupe-args' {
 }
 
 declare module 'eslint/lib/rules/no-dupe-class-members' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-dupe-else-if' {
   declare module.exports: any;
 }
 
@@ -674,6 +746,10 @@ declare module 'eslint/lib/rules/no-implied-eval' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-import-assign' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-inline-comments' {
   declare module.exports: any;
 }
@@ -718,7 +794,15 @@ declare module 'eslint/lib/rules/no-loop-func' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-loss-of-precision' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-magic-numbers' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-misleading-character-class' {
   declare module.exports: any;
 }
 
@@ -790,6 +874,10 @@ declare module 'eslint/lib/rules/no-new' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-nonoctal-decimal-escape' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-obj-calls' {
   declare module.exports: any;
 }
@@ -822,6 +910,10 @@ declare module 'eslint/lib/rules/no-process-exit' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-promise-executor-return' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-proto' {
   declare module.exports: any;
 }
@@ -835,6 +927,10 @@ declare module 'eslint/lib/rules/no-redeclare' {
 }
 
 declare module 'eslint/lib/rules/no-regex-spaces' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-restricted-exports' {
   declare module.exports: any;
 }
 
@@ -879,6 +975,10 @@ declare module 'eslint/lib/rules/no-self-compare' {
 }
 
 declare module 'eslint/lib/rules/no-sequences' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-setter-return' {
   declare module.exports: any;
 }
 
@@ -954,6 +1054,10 @@ declare module 'eslint/lib/rules/no-unneeded-ternary' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-unreachable-loop' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-unreachable' {
   declare module.exports: any;
 }
@@ -966,11 +1070,19 @@ declare module 'eslint/lib/rules/no-unsafe-negation' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-unsafe-optional-chaining' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-unused-expressions' {
   declare module.exports: any;
 }
 
 declare module 'eslint/lib/rules/no-unused-labels' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-unused-private-class-members' {
   declare module.exports: any;
 }
 
@@ -982,7 +1094,15 @@ declare module 'eslint/lib/rules/no-use-before-define' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/no-useless-backreference' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/no-useless-call' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-useless-catch' {
   declare module.exports: any;
 }
 
@@ -1086,7 +1206,23 @@ declare module 'eslint/lib/rules/prefer-destructuring' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/prefer-exponentiation-operator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/prefer-named-capture-group' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/prefer-numeric-literals' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/prefer-object-has-own' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/prefer-object-spread' {
   declare module.exports: any;
 }
 
@@ -1095,6 +1231,10 @@ declare module 'eslint/lib/rules/prefer-promise-reject-errors' {
 }
 
 declare module 'eslint/lib/rules/prefer-reflect' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/prefer-regex-literals' {
   declare module.exports: any;
 }
 
@@ -1122,11 +1262,19 @@ declare module 'eslint/lib/rules/radix' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/require-atomic-updates' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/require-await' {
   declare module.exports: any;
 }
 
 declare module 'eslint/lib/rules/require-jsdoc' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/require-unicode-regexp' {
   declare module.exports: any;
 }
 
@@ -1214,6 +1362,46 @@ declare module 'eslint/lib/rules/use-isnan' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/utils/ast-utils' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/fix-tracker' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/keywords' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/lazy-loading-rule-map' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/patterns/letters' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-combining-character' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-emoji-modifier' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-regional-indicator-symbol' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-surrogate-pair' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/valid-jsdoc' {
   declare module.exports: any;
 }
@@ -1242,147 +1430,147 @@ declare module 'eslint/lib/rules/yoda' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/testers/rule-tester' {
+declare module 'eslint/lib/shared/ajv' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/timing' {
+declare module 'eslint/lib/shared/ast-utils' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/backward-token-comment-cursor' {
+declare module 'eslint/lib/shared/config-validator' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/backward-token-cursor' {
+declare module 'eslint/lib/shared/deprecation-warnings' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/cursor' {
+declare module 'eslint/lib/shared/logging' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/cursors' {
+declare module 'eslint/lib/shared/relative-module-resolver' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/decorative-cursor' {
+declare module 'eslint/lib/shared/runtime-info' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/filter-cursor' {
+declare module 'eslint/lib/shared/string-utils' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/forward-token-comment-cursor' {
+declare module 'eslint/lib/shared/traverser' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/forward-token-cursor' {
+declare module 'eslint/lib/shared/types' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/index' {
+declare module 'eslint/lib/source-code' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/limit-cursor' {
+declare module 'eslint/lib/source-code/source-code' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/padded-token-cursor' {
+declare module 'eslint/lib/source-code/token-store/backward-token-comment-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/skip-cursor' {
+declare module 'eslint/lib/source-code/token-store/backward-token-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/utils' {
+declare module 'eslint/lib/source-code/token-store/cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/ajv' {
+declare module 'eslint/lib/source-code/token-store/cursors' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/apply-disable-directives' {
+declare module 'eslint/lib/source-code/token-store/decorative-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/fix-tracker' {
+declare module 'eslint/lib/source-code/token-store/filter-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/glob-util' {
+declare module 'eslint/lib/source-code/token-store/forward-token-comment-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/glob' {
+declare module 'eslint/lib/source-code/token-store/forward-token-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/hash' {
+declare module 'eslint/lib/source-code/token-store' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/interpolate' {
+declare module 'eslint/lib/source-code/token-store/limit-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/keywords' {
+declare module 'eslint/lib/source-code/token-store/padded-token-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/module-resolver' {
+declare module 'eslint/lib/source-code/token-store/skip-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/naming' {
+declare module 'eslint/lib/source-code/token-store/utils' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/node-event-generator' {
+declare module 'eslint/lib/unsupported-api' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/npm-util' {
+declare module 'eslint/messages/all-files-ignored' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/path-util' {
+declare module 'eslint/messages/extend-config-missing' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/patterns/letters' {
+declare module 'eslint/messages/failed-to-read-json' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/rule-fixer' {
+declare module 'eslint/messages/file-not-found' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/safe-emitter' {
+declare module 'eslint/messages/no-config-found' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/source-code-fixer' {
+declare module 'eslint/messages/plugin-conflict' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/source-code-util' {
+declare module 'eslint/messages/plugin-invalid' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/source-code' {
+declare module 'eslint/messages/plugin-missing' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/traverser' {
+declare module 'eslint/messages/print-config-with-directory-path' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/xml-escape' {
+declare module 'eslint/messages/whitespace-found' {
   declare module.exports: any;
 }
 
@@ -1396,140 +1584,173 @@ declare module 'eslint/conf/config-schema.js' {
 declare module 'eslint/conf/default-cli-options.js' {
   declare module.exports: $Exports<'eslint/conf/default-cli-options'>;
 }
-declare module 'eslint/conf/default-config-options.js' {
-  declare module.exports: $Exports<'eslint/conf/default-config-options'>;
-}
-declare module 'eslint/conf/environments.js' {
-  declare module.exports: $Exports<'eslint/conf/environments'>;
-}
 declare module 'eslint/conf/eslint-all.js' {
   declare module.exports: $Exports<'eslint/conf/eslint-all'>;
 }
 declare module 'eslint/conf/eslint-recommended.js' {
   declare module.exports: $Exports<'eslint/conf/eslint-recommended'>;
 }
+declare module 'eslint/conf/globals.js' {
+  declare module.exports: $Exports<'eslint/conf/globals'>;
+}
 declare module 'eslint/lib/api.js' {
   declare module.exports: $Exports<'eslint/lib/api'>;
 }
-declare module 'eslint/lib/ast-utils.js' {
-  declare module.exports: $Exports<'eslint/lib/ast-utils'>;
+declare module 'eslint/lib/cli-engine/cli-engine.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/cli-engine'>;
 }
-declare module 'eslint/lib/cli-engine.js' {
+declare module 'eslint/lib/cli-engine/file-enumerator.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/file-enumerator'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/checkstyle.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/checkstyle'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/compact.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/compact'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/html.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/html'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/jslint-xml.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/jslint-xml'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/json-with-metadata.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/json-with-metadata'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/json.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/json'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/junit.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/junit'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/stylish.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/stylish'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/tap.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/tap'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/unix.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/unix'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/visualstudio.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/visualstudio'>;
+}
+declare module 'eslint/lib/cli-engine/hash.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/hash'>;
+}
+declare module 'eslint/lib/cli-engine/index' {
   declare module.exports: $Exports<'eslint/lib/cli-engine'>;
+}
+declare module 'eslint/lib/cli-engine/index.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine'>;
+}
+declare module 'eslint/lib/cli-engine/lint-result-cache.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/lint-result-cache'>;
+}
+declare module 'eslint/lib/cli-engine/load-rules.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/load-rules'>;
+}
+declare module 'eslint/lib/cli-engine/xml-escape.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/xml-escape'>;
 }
 declare module 'eslint/lib/cli.js' {
   declare module.exports: $Exports<'eslint/lib/cli'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path-analyzer.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path-analyzer'>;
+declare module 'eslint/lib/config/default-config.js' {
+  declare module.exports: $Exports<'eslint/lib/config/default-config'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path-segment.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path-segment'>;
+declare module 'eslint/lib/config/flat-config-array.js' {
+  declare module.exports: $Exports<'eslint/lib/config/flat-config-array'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path-state.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path-state'>;
+declare module 'eslint/lib/config/flat-config-helpers.js' {
+  declare module.exports: $Exports<'eslint/lib/config/flat-config-helpers'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path'>;
+declare module 'eslint/lib/config/flat-config-schema.js' {
+  declare module.exports: $Exports<'eslint/lib/config/flat-config-schema'>;
 }
-declare module 'eslint/lib/code-path-analysis/debug-helpers.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/debug-helpers'>;
+declare module 'eslint/lib/config/rule-validator.js' {
+  declare module.exports: $Exports<'eslint/lib/config/rule-validator'>;
 }
-declare module 'eslint/lib/code-path-analysis/fork-context.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/fork-context'>;
+declare module 'eslint/lib/eslint/eslint.js' {
+  declare module.exports: $Exports<'eslint/lib/eslint/eslint'>;
 }
-declare module 'eslint/lib/code-path-analysis/id-generator.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/id-generator'>;
+declare module 'eslint/lib/eslint/index' {
+  declare module.exports: $Exports<'eslint/lib/eslint'>;
 }
-declare module 'eslint/lib/config.js' {
-  declare module.exports: $Exports<'eslint/lib/config'>;
+declare module 'eslint/lib/eslint/index.js' {
+  declare module.exports: $Exports<'eslint/lib/eslint'>;
 }
-declare module 'eslint/lib/config/autoconfig.js' {
-  declare module.exports: $Exports<'eslint/lib/config/autoconfig'>;
+declare module 'eslint/lib/linter/apply-disable-directives.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/apply-disable-directives'>;
 }
-declare module 'eslint/lib/config/config-cache.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-cache'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path-analyzer.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/code-path-analyzer'>;
 }
-declare module 'eslint/lib/config/config-file.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-file'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path-segment.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/code-path-segment'>;
 }
-declare module 'eslint/lib/config/config-initializer.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-initializer'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path-state.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/code-path-state'>;
 }
-declare module 'eslint/lib/config/config-ops.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-ops'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/code-path'>;
 }
-declare module 'eslint/lib/config/config-rule.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-rule'>;
+declare module 'eslint/lib/linter/code-path-analysis/debug-helpers.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/debug-helpers'>;
 }
-declare module 'eslint/lib/config/config-validator.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-validator'>;
+declare module 'eslint/lib/linter/code-path-analysis/fork-context.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/fork-context'>;
 }
-declare module 'eslint/lib/config/environments.js' {
-  declare module.exports: $Exports<'eslint/lib/config/environments'>;
+declare module 'eslint/lib/linter/code-path-analysis/id-generator.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/code-path-analysis/id-generator'>;
 }
-declare module 'eslint/lib/config/plugins.js' {
-  declare module.exports: $Exports<'eslint/lib/config/plugins'>;
+declare module 'eslint/lib/linter/config-comment-parser.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/config-comment-parser'>;
 }
-declare module 'eslint/lib/file-finder.js' {
-  declare module.exports: $Exports<'eslint/lib/file-finder'>;
-}
-declare module 'eslint/lib/formatters/checkstyle.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/checkstyle'>;
-}
-declare module 'eslint/lib/formatters/codeframe.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/codeframe'>;
-}
-declare module 'eslint/lib/formatters/compact.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/compact'>;
-}
-declare module 'eslint/lib/formatters/html.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/html'>;
-}
-declare module 'eslint/lib/formatters/jslint-xml.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/jslint-xml'>;
-}
-declare module 'eslint/lib/formatters/json.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/json'>;
-}
-declare module 'eslint/lib/formatters/junit.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/junit'>;
-}
-declare module 'eslint/lib/formatters/stylish.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/stylish'>;
-}
-declare module 'eslint/lib/formatters/table.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/table'>;
-}
-declare module 'eslint/lib/formatters/tap.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/tap'>;
-}
-declare module 'eslint/lib/formatters/unix.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/unix'>;
-}
-declare module 'eslint/lib/formatters/visualstudio.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/visualstudio'>;
-}
-declare module 'eslint/lib/ignored-paths.js' {
-  declare module.exports: $Exports<'eslint/lib/ignored-paths'>;
-}
-declare module 'eslint/lib/linter.js' {
+declare module 'eslint/lib/linter/index' {
   declare module.exports: $Exports<'eslint/lib/linter'>;
 }
-declare module 'eslint/lib/load-rules.js' {
-  declare module.exports: $Exports<'eslint/lib/load-rules'>;
+declare module 'eslint/lib/linter/index.js' {
+  declare module.exports: $Exports<'eslint/lib/linter'>;
 }
-declare module 'eslint/lib/logging.js' {
-  declare module.exports: $Exports<'eslint/lib/logging'>;
+declare module 'eslint/lib/linter/interpolate.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/interpolate'>;
+}
+declare module 'eslint/lib/linter/linter.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/linter'>;
+}
+declare module 'eslint/lib/linter/node-event-generator.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/node-event-generator'>;
+}
+declare module 'eslint/lib/linter/report-translator.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/report-translator'>;
+}
+declare module 'eslint/lib/linter/rule-fixer.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/rule-fixer'>;
+}
+declare module 'eslint/lib/linter/rules.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/rules'>;
+}
+declare module 'eslint/lib/linter/safe-emitter.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/safe-emitter'>;
+}
+declare module 'eslint/lib/linter/source-code-fixer.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/source-code-fixer'>;
+}
+declare module 'eslint/lib/linter/timing.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/timing'>;
 }
 declare module 'eslint/lib/options.js' {
   declare module.exports: $Exports<'eslint/lib/options'>;
 }
-declare module 'eslint/lib/report-translator.js' {
-  declare module.exports: $Exports<'eslint/lib/report-translator'>;
+declare module 'eslint/lib/rule-tester/index' {
+  declare module.exports: $Exports<'eslint/lib/rule-tester'>;
 }
-declare module 'eslint/lib/rules.js' {
-  declare module.exports: $Exports<'eslint/lib/rules'>;
+declare module 'eslint/lib/rule-tester/index.js' {
+  declare module.exports: $Exports<'eslint/lib/rule-tester'>;
+}
+declare module 'eslint/lib/rule-tester/rule-tester.js' {
+  declare module.exports: $Exports<'eslint/lib/rule-tester/rule-tester'>;
 }
 declare module 'eslint/lib/rules/accessor-pairs.js' {
   declare module.exports: $Exports<'eslint/lib/rules/accessor-pairs'>;
@@ -1603,8 +1824,14 @@ declare module 'eslint/lib/rules/constructor-super.js' {
 declare module 'eslint/lib/rules/curly.js' {
   declare module.exports: $Exports<'eslint/lib/rules/curly'>;
 }
+declare module 'eslint/lib/rules/default-case-last.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/default-case-last'>;
+}
 declare module 'eslint/lib/rules/default-case.js' {
   declare module.exports: $Exports<'eslint/lib/rules/default-case'>;
+}
+declare module 'eslint/lib/rules/default-param-last.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/default-param-last'>;
 }
 declare module 'eslint/lib/rules/dot-location.js' {
   declare module.exports: $Exports<'eslint/lib/rules/dot-location'>;
@@ -1633,6 +1860,9 @@ declare module 'eslint/lib/rules/func-names.js' {
 declare module 'eslint/lib/rules/func-style.js' {
   declare module.exports: $Exports<'eslint/lib/rules/func-style'>;
 }
+declare module 'eslint/lib/rules/function-call-argument-newline.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/function-call-argument-newline'>;
+}
 declare module 'eslint/lib/rules/function-paren-newline.js' {
   declare module.exports: $Exports<'eslint/lib/rules/function-paren-newline'>;
 }
@@ -1645,6 +1875,9 @@ declare module 'eslint/lib/rules/getter-return.js' {
 declare module 'eslint/lib/rules/global-require.js' {
   declare module.exports: $Exports<'eslint/lib/rules/global-require'>;
 }
+declare module 'eslint/lib/rules/grouped-accessor-pairs.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/grouped-accessor-pairs'>;
+}
 declare module 'eslint/lib/rules/guard-for-in.js' {
   declare module.exports: $Exports<'eslint/lib/rules/guard-for-in'>;
 }
@@ -1653,6 +1886,9 @@ declare module 'eslint/lib/rules/handle-callback-err.js' {
 }
 declare module 'eslint/lib/rules/id-blacklist.js' {
   declare module.exports: $Exports<'eslint/lib/rules/id-blacklist'>;
+}
+declare module 'eslint/lib/rules/id-denylist.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/id-denylist'>;
 }
 declare module 'eslint/lib/rules/id-length.js' {
   declare module.exports: $Exports<'eslint/lib/rules/id-length'>;
@@ -1668,6 +1904,12 @@ declare module 'eslint/lib/rules/indent-legacy.js' {
 }
 declare module 'eslint/lib/rules/indent.js' {
   declare module.exports: $Exports<'eslint/lib/rules/indent'>;
+}
+declare module 'eslint/lib/rules/index' {
+  declare module.exports: $Exports<'eslint/lib/rules'>;
+}
+declare module 'eslint/lib/rules/index.js' {
+  declare module.exports: $Exports<'eslint/lib/rules'>;
 }
 declare module 'eslint/lib/rules/init-declarations.js' {
   declare module.exports: $Exports<'eslint/lib/rules/init-declarations'>;
@@ -1696,11 +1938,17 @@ declare module 'eslint/lib/rules/lines-around-directive.js' {
 declare module 'eslint/lib/rules/lines-between-class-members.js' {
   declare module.exports: $Exports<'eslint/lib/rules/lines-between-class-members'>;
 }
+declare module 'eslint/lib/rules/max-classes-per-file.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/max-classes-per-file'>;
+}
 declare module 'eslint/lib/rules/max-depth.js' {
   declare module.exports: $Exports<'eslint/lib/rules/max-depth'>;
 }
 declare module 'eslint/lib/rules/max-len.js' {
   declare module.exports: $Exports<'eslint/lib/rules/max-len'>;
+}
+declare module 'eslint/lib/rules/max-lines-per-function.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/max-lines-per-function'>;
 }
 declare module 'eslint/lib/rules/max-lines.js' {
   declare module.exports: $Exports<'eslint/lib/rules/max-lines'>;
@@ -1744,6 +1992,9 @@ declare module 'eslint/lib/rules/no-alert.js' {
 declare module 'eslint/lib/rules/no-array-constructor.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-array-constructor'>;
 }
+declare module 'eslint/lib/rules/no-async-promise-executor.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-async-promise-executor'>;
+}
 declare module 'eslint/lib/rules/no-await-in-loop.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-await-in-loop'>;
 }
@@ -1783,6 +2034,9 @@ declare module 'eslint/lib/rules/no-const-assign.js' {
 declare module 'eslint/lib/rules/no-constant-condition.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-constant-condition'>;
 }
+declare module 'eslint/lib/rules/no-constructor-return.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-constructor-return'>;
+}
 declare module 'eslint/lib/rules/no-continue.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-continue'>;
 }
@@ -1803,6 +2057,9 @@ declare module 'eslint/lib/rules/no-dupe-args.js' {
 }
 declare module 'eslint/lib/rules/no-dupe-class-members.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-dupe-class-members'>;
+}
+declare module 'eslint/lib/rules/no-dupe-else-if.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-dupe-else-if'>;
 }
 declare module 'eslint/lib/rules/no-dupe-keys.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-dupe-keys'>;
@@ -1876,6 +2133,9 @@ declare module 'eslint/lib/rules/no-implicit-globals.js' {
 declare module 'eslint/lib/rules/no-implied-eval.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-implied-eval'>;
 }
+declare module 'eslint/lib/rules/no-import-assign.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-import-assign'>;
+}
 declare module 'eslint/lib/rules/no-inline-comments.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-inline-comments'>;
 }
@@ -1909,8 +2169,14 @@ declare module 'eslint/lib/rules/no-lonely-if.js' {
 declare module 'eslint/lib/rules/no-loop-func.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-loop-func'>;
 }
+declare module 'eslint/lib/rules/no-loss-of-precision.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-loss-of-precision'>;
+}
 declare module 'eslint/lib/rules/no-magic-numbers.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-magic-numbers'>;
+}
+declare module 'eslint/lib/rules/no-misleading-character-class.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-misleading-character-class'>;
 }
 declare module 'eslint/lib/rules/no-mixed-operators.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-mixed-operators'>;
@@ -1963,6 +2229,9 @@ declare module 'eslint/lib/rules/no-new-wrappers.js' {
 declare module 'eslint/lib/rules/no-new.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-new'>;
 }
+declare module 'eslint/lib/rules/no-nonoctal-decimal-escape.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-nonoctal-decimal-escape'>;
+}
 declare module 'eslint/lib/rules/no-obj-calls.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-obj-calls'>;
 }
@@ -1987,6 +2256,9 @@ declare module 'eslint/lib/rules/no-process-env.js' {
 declare module 'eslint/lib/rules/no-process-exit.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-process-exit'>;
 }
+declare module 'eslint/lib/rules/no-promise-executor-return.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-promise-executor-return'>;
+}
 declare module 'eslint/lib/rules/no-proto.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-proto'>;
 }
@@ -1998,6 +2270,9 @@ declare module 'eslint/lib/rules/no-redeclare.js' {
 }
 declare module 'eslint/lib/rules/no-regex-spaces.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-regex-spaces'>;
+}
+declare module 'eslint/lib/rules/no-restricted-exports.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-restricted-exports'>;
 }
 declare module 'eslint/lib/rules/no-restricted-globals.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-restricted-globals'>;
@@ -2031,6 +2306,9 @@ declare module 'eslint/lib/rules/no-self-compare.js' {
 }
 declare module 'eslint/lib/rules/no-sequences.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-sequences'>;
+}
+declare module 'eslint/lib/rules/no-setter-return.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-setter-return'>;
 }
 declare module 'eslint/lib/rules/no-shadow-restricted-names.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-shadow-restricted-names'>;
@@ -2086,6 +2364,9 @@ declare module 'eslint/lib/rules/no-unmodified-loop-condition.js' {
 declare module 'eslint/lib/rules/no-unneeded-ternary.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unneeded-ternary'>;
 }
+declare module 'eslint/lib/rules/no-unreachable-loop.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-unreachable-loop'>;
+}
 declare module 'eslint/lib/rules/no-unreachable.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unreachable'>;
 }
@@ -2095,11 +2376,17 @@ declare module 'eslint/lib/rules/no-unsafe-finally.js' {
 declare module 'eslint/lib/rules/no-unsafe-negation.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unsafe-negation'>;
 }
+declare module 'eslint/lib/rules/no-unsafe-optional-chaining.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-unsafe-optional-chaining'>;
+}
 declare module 'eslint/lib/rules/no-unused-expressions.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unused-expressions'>;
 }
 declare module 'eslint/lib/rules/no-unused-labels.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unused-labels'>;
+}
+declare module 'eslint/lib/rules/no-unused-private-class-members.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-unused-private-class-members'>;
 }
 declare module 'eslint/lib/rules/no-unused-vars.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unused-vars'>;
@@ -2107,8 +2394,14 @@ declare module 'eslint/lib/rules/no-unused-vars.js' {
 declare module 'eslint/lib/rules/no-use-before-define.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-use-before-define'>;
 }
+declare module 'eslint/lib/rules/no-useless-backreference.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-useless-backreference'>;
+}
 declare module 'eslint/lib/rules/no-useless-call.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-useless-call'>;
+}
+declare module 'eslint/lib/rules/no-useless-catch.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-useless-catch'>;
 }
 declare module 'eslint/lib/rules/no-useless-computed-key.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-useless-computed-key'>;
@@ -2185,14 +2478,29 @@ declare module 'eslint/lib/rules/prefer-const.js' {
 declare module 'eslint/lib/rules/prefer-destructuring.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-destructuring'>;
 }
+declare module 'eslint/lib/rules/prefer-exponentiation-operator.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/prefer-exponentiation-operator'>;
+}
+declare module 'eslint/lib/rules/prefer-named-capture-group.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/prefer-named-capture-group'>;
+}
 declare module 'eslint/lib/rules/prefer-numeric-literals.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-numeric-literals'>;
+}
+declare module 'eslint/lib/rules/prefer-object-has-own.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/prefer-object-has-own'>;
+}
+declare module 'eslint/lib/rules/prefer-object-spread.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/prefer-object-spread'>;
 }
 declare module 'eslint/lib/rules/prefer-promise-reject-errors.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-promise-reject-errors'>;
 }
 declare module 'eslint/lib/rules/prefer-reflect.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-reflect'>;
+}
+declare module 'eslint/lib/rules/prefer-regex-literals.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/prefer-regex-literals'>;
 }
 declare module 'eslint/lib/rules/prefer-rest-params.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-rest-params'>;
@@ -2212,11 +2520,17 @@ declare module 'eslint/lib/rules/quotes.js' {
 declare module 'eslint/lib/rules/radix.js' {
   declare module.exports: $Exports<'eslint/lib/rules/radix'>;
 }
+declare module 'eslint/lib/rules/require-atomic-updates.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/require-atomic-updates'>;
+}
 declare module 'eslint/lib/rules/require-await.js' {
   declare module.exports: $Exports<'eslint/lib/rules/require-await'>;
 }
 declare module 'eslint/lib/rules/require-jsdoc.js' {
   declare module.exports: $Exports<'eslint/lib/rules/require-jsdoc'>;
+}
+declare module 'eslint/lib/rules/require-unicode-regexp.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/require-unicode-regexp'>;
 }
 declare module 'eslint/lib/rules/require-yield.js' {
   declare module.exports: $Exports<'eslint/lib/rules/require-yield'>;
@@ -2281,6 +2595,39 @@ declare module 'eslint/lib/rules/unicode-bom.js' {
 declare module 'eslint/lib/rules/use-isnan.js' {
   declare module.exports: $Exports<'eslint/lib/rules/use-isnan'>;
 }
+declare module 'eslint/lib/rules/utils/ast-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/ast-utils'>;
+}
+declare module 'eslint/lib/rules/utils/fix-tracker.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/fix-tracker'>;
+}
+declare module 'eslint/lib/rules/utils/keywords.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/keywords'>;
+}
+declare module 'eslint/lib/rules/utils/lazy-loading-rule-map.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/lazy-loading-rule-map'>;
+}
+declare module 'eslint/lib/rules/utils/patterns/letters.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/patterns/letters'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/index' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/index.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-combining-character.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode/is-combining-character'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-emoji-modifier.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode/is-emoji-modifier'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-regional-indicator-symbol.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode/is-regional-indicator-symbol'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-surrogate-pair.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode/is-surrogate-pair'>;
+}
 declare module 'eslint/lib/rules/valid-jsdoc.js' {
   declare module.exports: $Exports<'eslint/lib/rules/valid-jsdoc'>;
 }
@@ -2302,111 +2649,117 @@ declare module 'eslint/lib/rules/yield-star-spacing.js' {
 declare module 'eslint/lib/rules/yoda.js' {
   declare module.exports: $Exports<'eslint/lib/rules/yoda'>;
 }
-declare module 'eslint/lib/testers/rule-tester.js' {
-  declare module.exports: $Exports<'eslint/lib/testers/rule-tester'>;
+declare module 'eslint/lib/shared/ajv.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/ajv'>;
 }
-declare module 'eslint/lib/timing.js' {
-  declare module.exports: $Exports<'eslint/lib/timing'>;
+declare module 'eslint/lib/shared/ast-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/ast-utils'>;
 }
-declare module 'eslint/lib/token-store/backward-token-comment-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/backward-token-comment-cursor'>;
+declare module 'eslint/lib/shared/config-validator.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/config-validator'>;
 }
-declare module 'eslint/lib/token-store/backward-token-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/backward-token-cursor'>;
+declare module 'eslint/lib/shared/deprecation-warnings.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/deprecation-warnings'>;
 }
-declare module 'eslint/lib/token-store/cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/cursor'>;
+declare module 'eslint/lib/shared/logging.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/logging'>;
 }
-declare module 'eslint/lib/token-store/cursors.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/cursors'>;
+declare module 'eslint/lib/shared/relative-module-resolver.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/relative-module-resolver'>;
 }
-declare module 'eslint/lib/token-store/decorative-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/decorative-cursor'>;
+declare module 'eslint/lib/shared/runtime-info.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/runtime-info'>;
 }
-declare module 'eslint/lib/token-store/filter-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/filter-cursor'>;
+declare module 'eslint/lib/shared/string-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/string-utils'>;
 }
-declare module 'eslint/lib/token-store/forward-token-comment-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/forward-token-comment-cursor'>;
+declare module 'eslint/lib/shared/traverser.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/traverser'>;
 }
-declare module 'eslint/lib/token-store/forward-token-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/forward-token-cursor'>;
+declare module 'eslint/lib/shared/types.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/types'>;
 }
-declare module 'eslint/lib/token-store/index.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/index'>;
+declare module 'eslint/lib/source-code/index' {
+  declare module.exports: $Exports<'eslint/lib/source-code'>;
 }
-declare module 'eslint/lib/token-store/limit-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/limit-cursor'>;
+declare module 'eslint/lib/source-code/index.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code'>;
 }
-declare module 'eslint/lib/token-store/padded-token-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/padded-token-cursor'>;
+declare module 'eslint/lib/source-code/source-code.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/source-code'>;
 }
-declare module 'eslint/lib/token-store/skip-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/skip-cursor'>;
+declare module 'eslint/lib/source-code/token-store/backward-token-comment-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/backward-token-comment-cursor'>;
 }
-declare module 'eslint/lib/token-store/utils.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/utils'>;
+declare module 'eslint/lib/source-code/token-store/backward-token-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/backward-token-cursor'>;
 }
-declare module 'eslint/lib/util/ajv.js' {
-  declare module.exports: $Exports<'eslint/lib/util/ajv'>;
+declare module 'eslint/lib/source-code/token-store/cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/cursor'>;
 }
-declare module 'eslint/lib/util/apply-disable-directives.js' {
-  declare module.exports: $Exports<'eslint/lib/util/apply-disable-directives'>;
+declare module 'eslint/lib/source-code/token-store/cursors.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/cursors'>;
 }
-declare module 'eslint/lib/util/fix-tracker.js' {
-  declare module.exports: $Exports<'eslint/lib/util/fix-tracker'>;
+declare module 'eslint/lib/source-code/token-store/decorative-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/decorative-cursor'>;
 }
-declare module 'eslint/lib/util/glob-util.js' {
-  declare module.exports: $Exports<'eslint/lib/util/glob-util'>;
+declare module 'eslint/lib/source-code/token-store/filter-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/filter-cursor'>;
 }
-declare module 'eslint/lib/util/glob.js' {
-  declare module.exports: $Exports<'eslint/lib/util/glob'>;
+declare module 'eslint/lib/source-code/token-store/forward-token-comment-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/forward-token-comment-cursor'>;
 }
-declare module 'eslint/lib/util/hash.js' {
-  declare module.exports: $Exports<'eslint/lib/util/hash'>;
+declare module 'eslint/lib/source-code/token-store/forward-token-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/forward-token-cursor'>;
 }
-declare module 'eslint/lib/util/interpolate.js' {
-  declare module.exports: $Exports<'eslint/lib/util/interpolate'>;
+declare module 'eslint/lib/source-code/token-store/index' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store'>;
 }
-declare module 'eslint/lib/util/keywords.js' {
-  declare module.exports: $Exports<'eslint/lib/util/keywords'>;
+declare module 'eslint/lib/source-code/token-store/index.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store'>;
 }
-declare module 'eslint/lib/util/module-resolver.js' {
-  declare module.exports: $Exports<'eslint/lib/util/module-resolver'>;
+declare module 'eslint/lib/source-code/token-store/limit-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/limit-cursor'>;
 }
-declare module 'eslint/lib/util/naming.js' {
-  declare module.exports: $Exports<'eslint/lib/util/naming'>;
+declare module 'eslint/lib/source-code/token-store/padded-token-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/padded-token-cursor'>;
 }
-declare module 'eslint/lib/util/node-event-generator.js' {
-  declare module.exports: $Exports<'eslint/lib/util/node-event-generator'>;
+declare module 'eslint/lib/source-code/token-store/skip-cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/skip-cursor'>;
 }
-declare module 'eslint/lib/util/npm-util.js' {
-  declare module.exports: $Exports<'eslint/lib/util/npm-util'>;
+declare module 'eslint/lib/source-code/token-store/utils.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/utils'>;
 }
-declare module 'eslint/lib/util/path-util.js' {
-  declare module.exports: $Exports<'eslint/lib/util/path-util'>;
+declare module 'eslint/lib/unsupported-api.js' {
+  declare module.exports: $Exports<'eslint/lib/unsupported-api'>;
 }
-declare module 'eslint/lib/util/patterns/letters.js' {
-  declare module.exports: $Exports<'eslint/lib/util/patterns/letters'>;
+declare module 'eslint/messages/all-files-ignored.js' {
+  declare module.exports: $Exports<'eslint/messages/all-files-ignored'>;
 }
-declare module 'eslint/lib/util/rule-fixer.js' {
-  declare module.exports: $Exports<'eslint/lib/util/rule-fixer'>;
+declare module 'eslint/messages/extend-config-missing.js' {
+  declare module.exports: $Exports<'eslint/messages/extend-config-missing'>;
 }
-declare module 'eslint/lib/util/safe-emitter.js' {
-  declare module.exports: $Exports<'eslint/lib/util/safe-emitter'>;
+declare module 'eslint/messages/failed-to-read-json.js' {
+  declare module.exports: $Exports<'eslint/messages/failed-to-read-json'>;
 }
-declare module 'eslint/lib/util/source-code-fixer.js' {
-  declare module.exports: $Exports<'eslint/lib/util/source-code-fixer'>;
+declare module 'eslint/messages/file-not-found.js' {
+  declare module.exports: $Exports<'eslint/messages/file-not-found'>;
 }
-declare module 'eslint/lib/util/source-code-util.js' {
-  declare module.exports: $Exports<'eslint/lib/util/source-code-util'>;
+declare module 'eslint/messages/no-config-found.js' {
+  declare module.exports: $Exports<'eslint/messages/no-config-found'>;
 }
-declare module 'eslint/lib/util/source-code.js' {
-  declare module.exports: $Exports<'eslint/lib/util/source-code'>;
+declare module 'eslint/messages/plugin-conflict.js' {
+  declare module.exports: $Exports<'eslint/messages/plugin-conflict'>;
 }
-declare module 'eslint/lib/util/traverser.js' {
-  declare module.exports: $Exports<'eslint/lib/util/traverser'>;
+declare module 'eslint/messages/plugin-invalid.js' {
+  declare module.exports: $Exports<'eslint/messages/plugin-invalid'>;
 }
-declare module 'eslint/lib/util/xml-escape.js' {
-  declare module.exports: $Exports<'eslint/lib/util/xml-escape'>;
+declare module 'eslint/messages/plugin-missing.js' {
+  declare module.exports: $Exports<'eslint/messages/plugin-missing'>;
+}
+declare module 'eslint/messages/print-config-with-directory-path.js' {
+  declare module.exports: $Exports<'eslint/messages/print-config-with-directory-path'>;
+}
+declare module 'eslint/messages/whitespace-found.js' {
+  declare module.exports: $Exports<'eslint/messages/whitespace-found'>;
 }
