@@ -1,5 +1,5 @@
-// flow-typed signature: 3b0221b63c706dd2a52cb9ae03e0da51
-// flow-typed version: <<STUB>>/expect_v^22.1.0/flow_v0.63.1
+// flow-typed signature: f9042e5070ff7cf1b186edab855bcf47
+// flow-typed version: <<STUB>>/expect_v^27.4.6/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,31 +22,23 @@ declare module 'expect' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'expect/build-es5/index' {
+declare module 'expect/build/asymmetricMatchers' {
   declare module.exports: any;
 }
 
-declare module 'expect/build/asymmetric_matchers' {
+declare module 'expect/build/extractExpectedAssertionsErrors' {
   declare module.exports: any;
 }
 
-declare module 'expect/build/extract_expected_assertions_errors' {
+declare module 'expect/build' {
   declare module.exports: any;
 }
 
-declare module 'expect/build/fake_chalk' {
+declare module 'expect/build/jasmineUtils' {
   declare module.exports: any;
 }
 
-declare module 'expect/build/index' {
-  declare module.exports: any;
-}
-
-declare module 'expect/build/jasmine_utils' {
-  declare module.exports: any;
-}
-
-declare module 'expect/build/jest_matchers_object' {
+declare module 'expect/build/jestMatchersObject' {
   declare module.exports: any;
 }
 
@@ -54,11 +46,19 @@ declare module 'expect/build/matchers' {
   declare module.exports: any;
 }
 
-declare module 'expect/build/spy_matchers' {
+declare module 'expect/build/print' {
   declare module.exports: any;
 }
 
-declare module 'expect/build/to_throw_matchers' {
+declare module 'expect/build/spyMatchers' {
+  declare module.exports: any;
+}
+
+declare module 'expect/build/toThrowMatchers' {
+  declare module.exports: any;
+}
+
+declare module 'expect/build/types' {
   declare module.exports: any;
 }
 
@@ -67,35 +67,38 @@ declare module 'expect/build/utils' {
 }
 
 // Filename aliases
-declare module 'expect/build-es5/index.js' {
-  declare module.exports: $Exports<'expect/build-es5/index'>;
+declare module 'expect/build/asymmetricMatchers.js' {
+  declare module.exports: $Exports<'expect/build/asymmetricMatchers'>;
 }
-declare module 'expect/build/asymmetric_matchers.js' {
-  declare module.exports: $Exports<'expect/build/asymmetric_matchers'>;
+declare module 'expect/build/extractExpectedAssertionsErrors.js' {
+  declare module.exports: $Exports<'expect/build/extractExpectedAssertionsErrors'>;
 }
-declare module 'expect/build/extract_expected_assertions_errors.js' {
-  declare module.exports: $Exports<'expect/build/extract_expected_assertions_errors'>;
-}
-declare module 'expect/build/fake_chalk.js' {
-  declare module.exports: $Exports<'expect/build/fake_chalk'>;
+declare module 'expect/build/index' {
+  declare module.exports: $Exports<'expect/build'>;
 }
 declare module 'expect/build/index.js' {
-  declare module.exports: $Exports<'expect/build/index'>;
+  declare module.exports: $Exports<'expect/build'>;
 }
-declare module 'expect/build/jasmine_utils.js' {
-  declare module.exports: $Exports<'expect/build/jasmine_utils'>;
+declare module 'expect/build/jasmineUtils.js' {
+  declare module.exports: $Exports<'expect/build/jasmineUtils'>;
 }
-declare module 'expect/build/jest_matchers_object.js' {
-  declare module.exports: $Exports<'expect/build/jest_matchers_object'>;
+declare module 'expect/build/jestMatchersObject.js' {
+  declare module.exports: $Exports<'expect/build/jestMatchersObject'>;
 }
 declare module 'expect/build/matchers.js' {
   declare module.exports: $Exports<'expect/build/matchers'>;
 }
-declare module 'expect/build/spy_matchers.js' {
-  declare module.exports: $Exports<'expect/build/spy_matchers'>;
+declare module 'expect/build/print.js' {
+  declare module.exports: $Exports<'expect/build/print'>;
 }
-declare module 'expect/build/to_throw_matchers.js' {
-  declare module.exports: $Exports<'expect/build/to_throw_matchers'>;
+declare module 'expect/build/spyMatchers.js' {
+  declare module.exports: $Exports<'expect/build/spyMatchers'>;
+}
+declare module 'expect/build/toThrowMatchers.js' {
+  declare module.exports: $Exports<'expect/build/toThrowMatchers'>;
+}
+declare module 'expect/build/types.js' {
+  declare module.exports: $Exports<'expect/build/types'>;
 }
 declare module 'expect/build/utils.js' {
   declare module.exports: $Exports<'expect/build/utils'>;
