@@ -1,5 +1,5 @@
-// flow-typed signature: 3d6877951e44ddb9c12ec68a8957f094
-// flow-typed version: <<STUB>>/flow-remove-types_v^1.2.3/flow_v0.63.1
+// flow-typed signature: 824c939334484f7db98408a6d14befe5
+// flow-typed version: <<STUB>>/flow-remove-types_v^2.170.0/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,6 +22,10 @@ declare module 'flow-remove-types' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
+declare module 'flow-remove-types/jest' {
+  declare module.exports: any;
+}
+
 declare module 'flow-remove-types/register' {
   declare module.exports: any;
 }
@@ -32,6 +36,9 @@ declare module 'flow-remove-types/index' {
 }
 declare module 'flow-remove-types/index.js' {
   declare module.exports: $Exports<'flow-remove-types'>;
+}
+declare module 'flow-remove-types/jest.js' {
+  declare module.exports: $Exports<'flow-remove-types/jest'>;
 }
 declare module 'flow-remove-types/register.js' {
   declare module.exports: $Exports<'flow-remove-types/register'>;
