@@ -1,5 +1,5 @@
-// flow-typed signature: 18900275c28a96722c5b911866345ad4
-// flow-typed version: <<STUB>>/flow-typed_v^2.2.3/flow_v0.63.1
+// flow-typed signature: ab764913d60de7c5133533ea6fc07f30
+// flow-typed version: <<STUB>>/flow-typed_v^3.6.1/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -23,6 +23,10 @@ declare module 'flow-typed' {
  * needed.
  */
 declare module 'flow-typed/dist/cli' {
+  declare module.exports: any;
+}
+
+declare module 'flow-typed/dist/commands/create-def' {
   declare module.exports: any;
 }
 
@@ -51,10 +55,6 @@ declare module 'flow-typed/dist/commands/update' {
 }
 
 declare module 'flow-typed/dist/commands/validateDefs' {
-  declare module.exports: any;
-}
-
-declare module 'flow-typed/dist/commands/version' {
   declare module.exports: any;
 }
 
@@ -114,13 +114,16 @@ declare module 'flow-typed/dist/lib/stubUtils' {
   declare module.exports: any;
 }
 
-declare module 'flow-typed/dist/lib/validationErrors' {
+declare module 'flow-typed/dist/lib/ValidationError' {
   declare module.exports: any;
 }
 
 // Filename aliases
 declare module 'flow-typed/dist/cli.js' {
   declare module.exports: $Exports<'flow-typed/dist/cli'>;
+}
+declare module 'flow-typed/dist/commands/create-def.js' {
+  declare module.exports: $Exports<'flow-typed/dist/commands/create-def'>;
 }
 declare module 'flow-typed/dist/commands/create-stub.js' {
   declare module.exports: $Exports<'flow-typed/dist/commands/create-stub'>;
@@ -142,9 +145,6 @@ declare module 'flow-typed/dist/commands/update.js' {
 }
 declare module 'flow-typed/dist/commands/validateDefs.js' {
   declare module.exports: $Exports<'flow-typed/dist/commands/validateDefs'>;
-}
-declare module 'flow-typed/dist/commands/version.js' {
-  declare module.exports: $Exports<'flow-typed/dist/commands/version'>;
 }
 declare module 'flow-typed/dist/lib/cacheRepoUtils.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/cacheRepoUtils'>;
@@ -188,6 +188,6 @@ declare module 'flow-typed/dist/lib/semver.js' {
 declare module 'flow-typed/dist/lib/stubUtils.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/stubUtils'>;
 }
-declare module 'flow-typed/dist/lib/validationErrors.js' {
-  declare module.exports: $Exports<'flow-typed/dist/lib/validationErrors'>;
+declare module 'flow-typed/dist/lib/ValidationError.js' {
+  declare module.exports: $Exports<'flow-typed/dist/lib/ValidationError'>;
 }
