@@ -1,5 +1,5 @@
-// flow-typed signature: 6c448ffea0f6185b7bd8f1e87a86a294
-// flow-typed version: <<STUB>>/ipfsd-ctl_v^0.27.0/flow_v0.63.1
+// flow-typed signature: 6069647686d59761c1ec78a7bbab4395
+// flow-typed version: <<STUB>>/ipfsd-ctl_v^10.0.5/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,228 +22,91 @@ declare module 'ipfsd-ctl' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'ipfsd-ctl/dist/index' {
-  declare module.exports: any;
-}
-
 declare module 'ipfsd-ctl/dist/index.min' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/examples/electron-asar/app' {
+declare module 'ipfsd-ctl/src/config' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/examples/id/id' {
+declare module 'ipfsd-ctl/src/endpoint/routes' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/examples/local-disposable/local-disposable' {
+declare module 'ipfsd-ctl/src/endpoint/server.browser' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/examples/local/local' {
+declare module 'ipfsd-ctl/src/endpoint/server' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/examples/remote-disposable/remote-disposable' {
+declare module 'ipfsd-ctl/src/factory' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/src/daemon-ctrl' {
+declare module 'ipfsd-ctl/src' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/src/daemon-node' {
+declare module 'ipfsd-ctl/src/ipfsd-client' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/src/exec' {
+declare module 'ipfsd-ctl/src/ipfsd-daemon' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/src/in-proc-node' {
+declare module 'ipfsd-ctl/src/ipfsd-in-proc' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/src/index' {
+declare module 'ipfsd-ctl/src/utils.browser' {
   declare module.exports: any;
 }
 
-declare module 'ipfsd-ctl/src/remote-node/client' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/src/remote-node/index' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/src/remote-node/routes' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/src/remote-node/server' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/src/utils/create-repo-browser' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/src/utils/create-repo-nodejs' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/src/utils/index' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/add-retrive' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/api' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/browser' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/daemon' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/exec' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/node' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/npm-installs' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/remote/client' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/remote/routes' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/remote/server' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/spawning' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/start-stop' {
-  declare module.exports: any;
-}
-
-declare module 'ipfsd-ctl/test/utils' {
+declare module 'ipfsd-ctl/src/utils' {
   declare module.exports: any;
 }
 
 // Filename aliases
-declare module 'ipfsd-ctl/dist/index.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/dist/index'>;
-}
 declare module 'ipfsd-ctl/dist/index.min.js' {
   declare module.exports: $Exports<'ipfsd-ctl/dist/index.min'>;
 }
-declare module 'ipfsd-ctl/examples/electron-asar/app.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/examples/electron-asar/app'>;
+declare module 'ipfsd-ctl/src/config.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/config'>;
 }
-declare module 'ipfsd-ctl/examples/id/id.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/examples/id/id'>;
+declare module 'ipfsd-ctl/src/endpoint/routes.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/endpoint/routes'>;
 }
-declare module 'ipfsd-ctl/examples/local-disposable/local-disposable.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/examples/local-disposable/local-disposable'>;
+declare module 'ipfsd-ctl/src/endpoint/server.browser.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/endpoint/server.browser'>;
 }
-declare module 'ipfsd-ctl/examples/local/local.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/examples/local/local'>;
+declare module 'ipfsd-ctl/src/endpoint/server.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/endpoint/server'>;
 }
-declare module 'ipfsd-ctl/examples/remote-disposable/remote-disposable.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/examples/remote-disposable/remote-disposable'>;
+declare module 'ipfsd-ctl/src/factory.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/factory'>;
 }
-declare module 'ipfsd-ctl/src/daemon-ctrl.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/daemon-ctrl'>;
-}
-declare module 'ipfsd-ctl/src/daemon-node.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/daemon-node'>;
-}
-declare module 'ipfsd-ctl/src/exec.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/exec'>;
-}
-declare module 'ipfsd-ctl/src/in-proc-node.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/in-proc-node'>;
+declare module 'ipfsd-ctl/src/index' {
+  declare module.exports: $Exports<'ipfsd-ctl/src'>;
 }
 declare module 'ipfsd-ctl/src/index.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/index'>;
+  declare module.exports: $Exports<'ipfsd-ctl/src'>;
 }
-declare module 'ipfsd-ctl/src/remote-node/client.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/remote-node/client'>;
+declare module 'ipfsd-ctl/src/ipfsd-client.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/ipfsd-client'>;
 }
-declare module 'ipfsd-ctl/src/remote-node/index.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/remote-node/index'>;
+declare module 'ipfsd-ctl/src/ipfsd-daemon.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/ipfsd-daemon'>;
 }
-declare module 'ipfsd-ctl/src/remote-node/routes.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/remote-node/routes'>;
+declare module 'ipfsd-ctl/src/ipfsd-in-proc.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/ipfsd-in-proc'>;
 }
-declare module 'ipfsd-ctl/src/remote-node/server.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/remote-node/server'>;
+declare module 'ipfsd-ctl/src/utils.browser.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/utils.browser'>;
 }
-declare module 'ipfsd-ctl/src/utils/create-repo-browser.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/utils/create-repo-browser'>;
-}
-declare module 'ipfsd-ctl/src/utils/create-repo-nodejs.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/utils/create-repo-nodejs'>;
-}
-declare module 'ipfsd-ctl/src/utils/index.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/src/utils/index'>;
-}
-declare module 'ipfsd-ctl/test/add-retrive.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/add-retrive'>;
-}
-declare module 'ipfsd-ctl/test/api.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/api'>;
-}
-declare module 'ipfsd-ctl/test/browser.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/browser'>;
-}
-declare module 'ipfsd-ctl/test/daemon.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/daemon'>;
-}
-declare module 'ipfsd-ctl/test/exec.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/exec'>;
-}
-declare module 'ipfsd-ctl/test/node.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/node'>;
-}
-declare module 'ipfsd-ctl/test/npm-installs.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/npm-installs'>;
-}
-declare module 'ipfsd-ctl/test/remote/client.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/remote/client'>;
-}
-declare module 'ipfsd-ctl/test/remote/routes.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/remote/routes'>;
-}
-declare module 'ipfsd-ctl/test/remote/server.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/remote/server'>;
-}
-declare module 'ipfsd-ctl/test/spawning.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/spawning'>;
-}
-declare module 'ipfsd-ctl/test/start-stop.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/start-stop'>;
-}
-declare module 'ipfsd-ctl/test/utils.js' {
-  declare module.exports: $Exports<'ipfsd-ctl/test/utils'>;
+declare module 'ipfsd-ctl/src/utils.js' {
+  declare module.exports: $Exports<'ipfsd-ctl/src/utils'>;
 }
