@@ -1,5 +1,5 @@
-// flow-typed signature: 2fcfb40af9c087a63e331b79c82dd836
-// flow-typed version: <<STUB>>/json-stringify-deterministic_v^1.0.1/flow_v0.63.1
+// flow-typed signature: cb27506e6ae80f8be7ea57ff64d4e355
+// flow-typed version: <<STUB>>/json-stringify-deterministic_v^1.0.2/flow_v0.130.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -26,7 +26,7 @@ declare module 'json-stringify-deterministic/lib/defaults' {
   declare module.exports: any;
 }
 
-declare module 'json-stringify-deterministic/lib/index' {
+declare module 'json-stringify-deterministic/lib' {
   declare module.exports: any;
 }
 
@@ -38,8 +38,11 @@ declare module 'json-stringify-deterministic/lib/util' {
 declare module 'json-stringify-deterministic/lib/defaults.js' {
   declare module.exports: $Exports<'json-stringify-deterministic/lib/defaults'>;
 }
+declare module 'json-stringify-deterministic/lib/index' {
+  declare module.exports: $Exports<'json-stringify-deterministic/lib'>;
+}
 declare module 'json-stringify-deterministic/lib/index.js' {
-  declare module.exports: $Exports<'json-stringify-deterministic/lib/index'>;
+  declare module.exports: $Exports<'json-stringify-deterministic/lib'>;
 }
 declare module 'json-stringify-deterministic/lib/util.js' {
   declare module.exports: $Exports<'json-stringify-deterministic/lib/util'>;
